### 2020-01-21 #2033 @whedon
[PRE REVIEW]: FASMA 2.0: A Python package for stellar parameters and chemical abundances
**Submitting author:** @MariaTsantaki (<a href="http://orcid.org/0000-0002-0552-2313">Maria Tsantaki</a>)
**Repository:** <a href="https://github.com/MariaTsantaki/FASMA-synthesis" target ="_blank">https://github.com/MariaTsantaki/FASMA-synthesis</a>
**Version:** v2.0
