### 2020-01-23 #2033 @DanielAndreasen
Dear @xuanxu (or if another is going to be editor on this submission). First of all, thank you for your great work you do here.
In regard to the code, I would like to point out that MOOG (in `FASMA/MOOG`) is an external software. We have included it here (as similar software to FASMA has done too) for convenience of the end user.
Therefor I do not think that part requires much attention/review. Rather the python pipeline that is
FASMA. I'm saying this since MOOG is quite comlicated and will take a long time to go through and understand.

Do you agree @MariaTsantaki?
