### 2020-01-23 #2033 @xuanxu
@DanielAndreasen: about redistributing MOOG with FASMA: you are releasing FASMA under MIT so to avoid license conflicts I'd recommend to add MOOG's license file inside the `FASMA/MOOG` directory (I don't know what their license is) and mention somewhere you are just redistributing it without modifications.
