### 2020-01-23 #2033 @xuanxu
@MariaTsantaki whedon suggest [this](https://ui.adsabs.harvard.edu/abs/1974ApJ...189..493S/abstract) as the article referenced by your `Carbon and Nitrogen Abundances in Metal-Poor Stars` entry in the paper.bib file. Is that it? Author and title are the same but publication year is different.
