### 2020-01-27 #2033 @MariaTsantaki
> @MariaTsantaki whedon suggest [this](https://ui.adsabs.harvard.edu/abs/1974ApJ...189..493S/abstract) as the article referenced by your `Carbon and Nitrogen Abundances in Metal-Poor Stars` entry in the paper.bib file. Is that it? Author and title are the same but publication year is different.

Thanks! I was referring to the PhD thesis where the code was first presented but maybe the format of that reference is not valid for JOSS. I can change it to the reviewed journal reference.
