### 2020-01-27 68b9267
bib corrected, moog license added
