+++ paper.bib
@@ -1,10 +1,14 @@
-@PHDTHESIS{Sneden:1973,
-   author = {{Sneden}, C.~A.},
-    title = "{Carbon and Nitrogen Abundances in Metal-Poor Stars.}",
-   school = {The University of Texas at Austin.},
-     year = 1973,
-   adsurl = {http://adsabs.harvard.edu/abs/1973PhDT.......180S},
-  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+@ARTICLE{Sneden:1973,
+       author = {{Sneden}, C.},
+        title = "{The nitrogen abundance of the very metal-poor star HD 122563.}",
+      journal = {\apj},
+         year = "1973",
+        month = "Sep",
+       volume = {184},
+        pages = {839},
+          doi = {10.1086/152374},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1973ApJ...184..839S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @ARTICLE{Kurucz:1993,
