### 2020-01-29 #2033 @xuanxu
:wave: @ipelupessy: would you be willing to review this paper? I think your knowledge of astronomy and Python makes you a great fit for this submission.
