### 2020-01-29 #2033 @xuanxu
:wave: @warrickball would you be willing to review this paper? I think your expertise on stellar physics makes you a great fit for this submission.
