### 2020-01-29 #2033 @warrickball
Hi @xuanxu, I'm happy to review this paper. Let me know how to proceed.

It isn't clear to me how the bundling of an external program happens in JOSS. FASMA appears to be a wrapper for the Fortran 77 code MOOG, presumably to make it easier to use. Though the [license file for MOOG in the repo](https://github.com/MariaTsantaki/FASMA-synthesis/blob/68b9267aa53a1e7023f84f59a4dad76f815cfe3b/FASMA/MOOG/license.txt) says "MOOG has public permission and free of charge", the [link](https://www.as.utexas.edu/~chris/moog.html) doesn't have an obvious (to me, anyway) license file or statement.

I assume this is a common problem that we might not solve now. Is there any precedent for this situation in JOSS (i.e. are there packages in JOSS that are wrappers to external codes)? Perhaps the repo should *not* bundle MOOG but rather give instructions on how to download and install it?
