### 2020-01-29 #2033 @xuanxu
> I'm happy to review this paper. Let me know how to proceed.

Great, thanks @warrickball! Once I have the second reviewer confirmed I'll start a new issue where the actual review will take place.

 > Though the [license file for MOOG in the repo](https://github.com/MariaTsantaki/FASMA-synthesis/blob/68b9267aa53a1e7023f84f59a4dad76f815cfe3b/FASMA/MOOG/license.txt) says "MOOG has public permission and free of charge", the [link](https://www.as.utexas.edu/~chris/moog.html) doesn't have an obvious (to me, anyway) license file or statement.
>

You are right, this should be clarified during the review to make sure MOOG allows redistribution of the software.

> I assume this is a common problem that we might not solve now. Is there any precedent for this situation in JOSS (i.e. are there packages in JOSS that are wrappers to external codes)?

Wrappers are common but in general the external codes are not included.

>Perhaps the repo should _not_ bundle MOOG but rather give instructions on how to download and install it?

Yes, if redistribution is not allowed, this is the easiest solution.
