### 2020-01-29 #2048 @ipelupessy
@MariaTsantaki I think the install procedure should not require sudo
