### 2020-01-29 #2048 @ipelupessy
@MariaTsantaki I see that the install procedure calls for editing the source code of MOOG by hand and then build/installing this seperately; it would be nicer if the setup.py does this for you (both things). Also the model data extraction can also be put in the setup.py. fixing this would allow a ```pip install FASMA```
