### 2020-01-29 #2048 @ipelupessy
@MariaTsantaki the repository contains compiled .o and executable files of MOOG. I would suggest these be removed.
