### 2020-01-29 #2048 @DanielAndreasen
@ipelupessy (I'm co-author of FASMA). It is true that there currently only are two unit tests. We plan to add more continously, but it hasn't been a priority.
