### 2020-01-29 #2048 @MariaTsantaki
@ipelupessy Thanks for reviewing! It's true, only 2 tests. My reason is that errors can be checked by modifying the configuration file and running fasma. But I can include more tests if it is more efficient for the testing.
