### 2020-01-29 #2048 @ipelupessy
ok, you need at least basic tests where you go through some simple cases (confirming the install went ok etc) - this will help you enormously when you change something in the install, making sure nothing breaks..

another comment: have you tried to minimize your prerequisites? e.g. I think at the moment astropy is only needed for for fits support - its a big prereq. for a small task..your package is not too bad..but in general I am in favour of minimizing dpendencies;-)

I will try compile and run something later so I can comment more substantially...
