### 2020-01-29 MariaTsantaki/FASMA-synthesis#12 @warrickball
Failure to install MOOG on Fedora 31 following FASMA instructions
This is connected to the [review of this repo](https://github.com/openjournals/joss-reviews/issues/2048) on JOSS.

I've just tried to install MOOG, following the FASMA instructions, and it can't find some libraries that are probably related to `sm` (SuperMongo?).

Line 29 of `Moogsilent.f` is already
````Fortran
      machine = "pcl"
````
which I assume is correct given that I'm using Fedora 31. I then ran `make -f Makefile.rh64silent` and eventually got the error:
````
gfortran -ff2c -Wall Abfind.o Abpop.o Abunplot.o Batom.o Begin.o Binary.o Binplot.o Binplotprep.o Blankstring.o Blends.o Bmolec.o Boxit.o Calmod.o Cdcalc.o Chabund.o Cog.o Cogplot.o Cogsyn.o Correl.o Crosscorr.o Curve.o Damping.o Defcolor.o Discov.o Doflux.o Drawcurs.o Eqlib.o Ewfind.o Ewweighted.o Fakeline.o Findtic.o Finish.o Fluxplot.o Gammabark.o Getasci.o Getcount.o Getnum.o Getsyns.o Gridplo.o Gridsyn.o Infile.o Inlines.o Inmodel.o Invert.o Jexpint.o Lineinfo.o Lineabund.o Linlimit.o Makeplot.o Molquery.o Moogsilent.o Mydriver.o Nansi.o Nearly.o Number.o Obshead.o Oneline.o Opaccouls.o OpacHelium.o OpacHydrogen.o Opacit.o Opacmetals.o Opacscat.o Params.o Partfn.o Partnew.o Plotit.o Plotremember.o Pltabun.o Pltcog.o Pltflux.o Pltspec.o Pointcurs.o Prinfo.o Putasci.o Readobs.o Rinteg.o Setmols.o Smooth.o Specplot.o Stats.o Sunder.o Synpop.o Synspec.o Synth.o Tablepop.o Taukap.o Total.o Trudamp.o Ucalc.o Vargauss.o Vmacro.o Voigt.o Wavecalc.o Weedout.o Writenumber.o -o MOOGSILENT -Lfakelib -lX11 \
        -Lfakelib -lplotsub -ldevices -lutils
/usr/bin/ld: cannot find -lplotsub
/usr/bin/ld: cannot find -ldevices
/usr/bin/ld: cannot find -lutils
collect2: error: ld returned 1 exit status
make: *** [Makefile.rh64silent:47: MOOGSILENT] Error 1
````
I've played around a little with the Makefile and I'm pretty sure the missing libraries are associated with `sm`. i.e. if I remove `-L$(SMLIB) -lplotsub -ldevices -lutils` I get lots of errors of the form
````
/usr/bin/ld: Cogplot.f:(.text+0x5e3): undefined reference to `sm_ltype__'
````
 I've spent some time trying to remove the files with the `sm` calls (mostly `*plot.f`) but I haven't managed to get anything working.
