### 2020-01-29 #2048 @warrickball
@MariaTsantaki I've just tried to install the program following the instructions on the GitHub repo. Like @ipelupessy, it shouldn't be necessary to install with `sudo`, so I had a look at the top-level [`makefile`](https://github.com/MariaTsantaki/FASMA-synthesis/blob/68b9267aa53a1e7023f84f59a4dad76f815cfe3b/makefile), on which I have several comments.

First, this doesn't make use of any of Make's dependency handling and doesn't do anything more than a bash script, so if the script remains necessary I think you should just run it as a bash script. But the contents are so simple that it would be better off as simple installation instructions.

Secondly, the `cp` and subsequent `rm` commands are equivalent to `mv`, so you could just use that. In fact, it'd make more sense to me to do
````
cd FASMA/models
tar -zxvf apogee_kurucz.tar.gz
tar -zxvf marcs.tar.gz
cd ../..
````

With the `pip` install listed in an instructions page, the user can decide whether to use `pip install` with or without the `--user` flag or `sudo`.

I'm also having trouble installing MOOG but, since that's actually stopping me from progressing, I've [opened an issue in the repo itself](https://github.com/MariaTsantaki/FASMA-synthesis/issues/12).
