### 2020-01-29 #2048 @warrickball
> @MariaTsantaki I see that the install procedure calls for editing the source code of MOOG by hand and then build/installing this seperately; it would be nicer if the setup.py does this for you (both things). Also the model data extraction can also be put in the setup.py. fixing this would allow a `pip install FASMA`

I'm not familiar with everything that `setup.py` can do but if it can handle extracting the atmosphere tables, then I agree that's the best way to handle installation (which would override my remarks above).
