### 2020-01-30 #2048 @ipelupessy
@MariaTsantaki  as @warrickball remarked, the installation instructions are not complete. You need sm plotting package. I am not sure the plotting is needed; in any case it compiles if you add a .f file with stubs for the sm calls.

Then, when running fasma, MOOG fails because missing data (Barklem.dat), I pulled this from https://github.com/hmtabernero/StePar (there is also a stub library for sm there), after this it seems to work.
