### 2020-01-30 4e6ebe0
do not hide MOOG output
