### 2020-01-30 #2048 @ipelupessy
some comments on running:
- the configuration with StarMe_synth.cfg is not very intuitive; it seems to need a particular formatting; it would be better if you could provide these through python keyword arguments directly
- I made a cfg file with just the following:
```
# this is a header (all special characters will act as header/comments)
# linelist teff logg [M/H] vt vmac vsini options
giraffe_sun_arcturus_calib.lst
```
fasma then complaints about a missing intervals file. so i copied over intervals.lst. it still complained. Error reporting doesn't actually tell me which file was missing, added this in the error message (see PR https://github.com/MariaTsantaki/FASMA-synthesis/pull/13 ). turns out the default suggested by the documentation is not correct (intervals_hr10_15n.lst)
- after this MOOG, fasma complained about missing summary.out. It turns out MOOG was giving an error about missing data. By default the output of MOOG is hidden though, so you would not know. This is not good behaviour - you can hide the errors, but then you should catch and act on them, or just report the output of MOOG (PR https://github.com/MariaTsantaki/FASMA-synthesis/pull/14)
- with the data (Barklem.dat) installed in the proper location, fasma runs and finishes. The output
files synthresults.dat and synthresults_elements.dat are empty apart from the header. I am not sure
this is an error of some sort or expected behaviour...
- fasma seems to be made to run in a particular location; for this sort of tool you want to be able to run it in some seperate project folder...(I ran in a test folder, I had to link everything through). It would be better if the needed file locations are pulled from the installation dirs.
