### 2020-01-30 #2048 @ipelupessy
@MariaTsantaki some further comments:
for this kind of wrapping it is nice if you can continue from within python with the results - so like:
result = fasma(**kwargs)
with kwargs the options as keyword arguments (as mentioned above) and result a python object encapsulating the output with some metadata. Now you are going to have to parse data files in some obscure format...
