### 2020-01-30 #2048 @ipelupessy
https://github.com/MariaTsantaki/FASMA-synthesis/pull/15 contains the above mentioned stub for sm
