### 2020-02-03 f312e24
setup changes, moog without sm
