+++ check_moog.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-if ! command -v MOOGSILENT &> /dev/null ; then
-    echo Please install MOOGSILENT!
-    exit 1
-fi
-    echo MOOGSILENT is installed
