+++ FASMA/MOOG/Abpop.f
@@ -214,7 +214,8 @@ c     options concerning what is seen on the plot
          call blankstring (moditle)
          moditle(1:70) = popitle(1:70)
          moditle(57:80) = 'EW-POPULATION '
-         call pltabun
+         print *, "Sorry I cannot plot!"
+c         call pltabun
       endif
 
 
