+++ FASMA/MOOG/Cogsyn.f
@@ -88,8 +88,8 @@ c*****do the syntheses
          pecabund(iatom,1) = pecabund(iatom,1) + rwstep
          go to 20
       endif
-      call pltcog
-
+c      call pltcog
+      print *, "Sorry I cannot plot!"
          
 c*****finish
       call finish (0)
