+++ FASMA/MOOG/Doflux.f
@@ -68,8 +68,8 @@ c*****compute the flux curve
       write (nf2out,1001) wave, flux, waveinv, fluxlog
       wave = wave + step
       if (wave .le. sstop) go to 1
-      call pltflux
-
+c      call pltflux
+      print *, "Sorry I cannot plot!"
 
 c****end the computations
       call finish (0)
