### 2020-02-04 95faa26
moog runs with fasma from anywhere
