+++ FASMA/utils.py
@@ -497,12 +497,13 @@ def _run_moog(par='batch.par', driver='synth'):
     ------
       Run MOOG once in silent mode
     '''
+    path = os.path.dirname(os.path.abspath(__file__))
     if driver == 'abfind':
-        os.system('./MOOG/MOOGSILENT > /dev/null')
+        os.system(path + '/MOOG/./MOOGSILENT > /dev/null')
     elif driver == 'synth':
         with open('stupid.tmp', 'w') as f:
             f.writelines('batch.par\nq')
-        os.system('./MOOG/MOOGSILENT < stupid.tmp > /dev/null 2>&1')
+        os.system(path + '/MOOG/./MOOGSILENT < stupid.tmp > /dev/null 2>&1')
         os.remove('stupid.tmp')
 
 
