+++ FASMA/StarMe_synth.cfg
@@ -1,3 +1,3 @@
 # this is a header (all special characters will act as header/comments)
 # linelist teff logg [M/H] vt vmac vsini options
-/home/bla/FASMA-synthesis/FASMA/rawLinelist/giraffe_sun_arcturus_calib.lst 5777 4.44 0.0 1.0 3.21 1.9 observations:/home/bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits,resolution:110000,inter_file:/home/bla/FASMA-synthesis/FASMA/intervals.lst,minimize,refine,plot
+/bla/FASMA-synthesis/FASMA/rawLinelist/linelist.lst 5777 4.44 0.0 1.0 3.21 1.9 observations:/bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits,resolution:110000,inter_file:/bla/FASMA-synthesis/FASMA/rawLinelist/intervals.lst,minimize,refine,plot
