+++ FASMA/synthDriver.py
@@ -5,16 +5,15 @@
 from __future__ import division, print_function
 import logging
 import os
-import yaml
 from .utils import fun_moog_synth as func
 from .observations import read_obs_intervals, plot, snr
 from .minimization import MinimizeSynth, getMic, getMac
 from .synthetic import read_linelist, read_linelist_elem, save_synth_spec
 import time
+import yaml
 
-
-class synthMethod:
-    def __init__(self, cfgfile='StarMe_synth.cfg', overwrite=None):
+class fasma:
+    def __init__(self, cfgfile='StarMe_synth.cfg', overwrite=None, **kwargs):
         '''The function that glues everything together. A log file is created
         with the list of processes (captain.log).
 
@@ -34,6 +33,7 @@ class synthMethod:
         self.cfgfile = cfgfile
         self.overwrite = overwrite
         self.status = None
+        self.kwargs = kwargs
         # Setup of logger
         if os.path.isfile('captain.log'):  # Cleaning from previous runs
             os.remove('captain.log')
@@ -48,11 +48,133 @@ class synthMethod:
         if not os.path.isdir('/MOOG/data'):
             path = os.path.dirname(os.path.abspath(__file__))
             os.system('cp -r ' + path + '/MOOG/data .')
+        self.configure(**kwargs)
+        self.synthdriver()
+        self.result()
 
-    @classmethod
-    def create_config(cls, cfgfile='StarMe_synth.cfg', **kwargs):
+    def configure(cls, cfgfile='StarMe_synth.cfg', **kwargs):
         '''Create configuration file from kwargs.
         Otherwise set to default.
+        '''
+        defaults = {
+            'linelist': 'linelist.lst',
+            'teff': 5777,
+            'logg': 4.44,
+            'feh': 0.0,
+            'vt': 1.0,
+            'vmac': 3.21,
+            'vsini': 1.90,
+            'model': 'apogee_kurucz',
+            'MOOGv': 2014,
+            'save': False,
+            'element': False,
+            'fix_teff': False,
+            'fix_logg': False,
+            'fix_feh': False,
+            'fix_vt': False,
+            'fix_vmac': False,
+            'fix_vsini': False,
+            'plot': False,
+            'plot_res': False,
+            'damping': 1,
+            'step_wave': 0.01,
+            'step_flux': 3.0,
+            'minimize': False,
+            'refine': False,
+            'observations': False,
+            'intervals_file': 'intervals.lst',
+            'snr': None,
+            'resolution': None,
+            'limb': 0.6
+        }
+
+        defaults.update(kwargs)
+        dic = {}
+        dic['star'] = defaults
+        with open(cfgfile, 'w') as f:
+            yaml.dump(dic, f)
+
+    def _setup(self, line):
+        '''Do the setup with initial parameters and options.
+
+        Input
+        -----
+        line : list
+          Each line from the configuration file after being split at spaces
+          The format is spaced separated: linelist_file (params) (options)
+        '''
+        self.linelist = line.linelist
+        self._options(line)
+        self.initial = [self.options['teff'], self.options['logg'], self.options['feh'], self.options['vt'], self.options['vmac'], self.options['vsini']]
+
+    def _genStar(self):
+        '''A generator for the configuration file.
+        '''
+        with open(cfgfile, 'r') as stream:
+	           try:
+                   data = yaml.safe_load(stream)
+	           except yaml.YAMLError as exc:
+                   self.logger.error('Could not process this information, check input parameters.')
+                   break
+
+        for item, line in data.items():
+            self._setup(line)
+            yield self.initial, self.options
+
+    def _prepare(self):
+        '''Check if linelist exists and create the first synthetic spectrum with
+        initial parameters.
+        '''
+
[…]
