+++ FASMA/StarMe_synth.cfg
@@ -1,3 +1,133 @@
-# this is a header (all special characters will act as header/comments)
-# linelist teff logg [M/H] vt vmac vsini options
-/bla/FASMA-synthesis/FASMA/rawLinelist/linelist.lst 5777 4.44 0.0 1.0 3.21 1.9 observations:/bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits,resolution:110000,inter_file:/bla/FASMA-synthesis/FASMA/rawLinelist/intervals.lst,minimize,refine,plot
+# lol
+star1: 
+    teff: 5777 
+    logg: 4.44 
+    feh: 0.0 
+    vt: 1.0 
+    vmac: 3.21 
+    vsini: 1.9 
+    observations: /bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits 
+    resolution: 110000 
+    inter_file: /bla/FASMA-synthesis/FASMA/rawLinelist/intervals.lst 
+    minimize: True 
+    refine: True 
+    plot: True
+star2: 
+    teff: 5777 
+    logg: 4.44 
+    feh: 0.0 
+    vt: 1.0 
+    vmac: 3.21 
+    vsini: 1.9 
+    observations: /bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits 
+    resolution: 110000 
+    inter_file: /bla/FASMA-synthesis/FASMA/rawLinelist/intervals.lst 
+    minimize: True 
+    refine: True 
+    plot: True
+star3: 
+    teff: 5777 
+    logg: 4.44 
+    feh: 0.0 
+    vt: 1.0 
+    vmac: 3.21 
+    vsini: 1.9 
+    observations: /bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits 
+    resolution: 110000 
+    inter_file: /bla/FASMA-synthesis/FASMA/rawLinelist/intervals.lst 
+    minimize: True 
+    refine: True 
+    plot: True
+star4: 
+    teff: 5777 
+    logg: 4.44 
+    feh: 0.0 
+    vt: 1.0 
+    vmac: 3.21 
+    vsini: 1.9 
+    observations: /bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits 
+    resolution: 110000 
+    inter_file: /bla/FASMA-synthesis/FASMA/rawLinelist/intervals.lst 
+    minimize: True 
+    refine: True 
+    plot: True
+
+star5: 
+    teff: 5777 
+    logg: 4.44 
+    feh: 0.0 
+    vt: 1.0 
+    vmac: 3.21 
+    vsini: 1.9 
+    observations: /bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits 
+    resolution: 110000 
+    inter_file: /bla/FASMA-synthesis/FASMA/rawLinelist/intervals.lst 
+    minimize: True 
+    refine: True 
+    plot: True
+
+star6: 
+    teff: 5777 
+    logg: 4.44 
+    feh: 0.0 
+    vt: 1.0 
+    vmac: 3.21 
+    vsini: 1.9 
+    observations: /bla/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits 
+    resolution: 110000 
+    inter_file: /bla/FASMA-synthesis/FASMA/rawLinelist/intervals.lst 
+    minimize: True 
+    refine: True 
+    plot: True
+star7: 
+    teff: 5777 
+    logg: 4.44 
+    feh: 0.0 
+    vt: 1.0 
+    vmac: 3.21 
+    vsini: 1.9 
[…]
