### 2020-02-09 1459a4e
StarMe_synth.cfg -> config.yml; captain.log -> fasma.log
