+++ FASMA/__init__.py
@@ -6,6 +6,6 @@ if __name__ == '__main__':
     if len(sys.argv) > 1:
         cfgfile = sys.argv[1]
     else:
-        cfgfile = 'StarMe_synth.cfg'
+        cfgfile = 'config.yml'
     driver = synthMethod(cfgfile=cfgfile, overwrite=None)
     driver.synthdriver()
