+++ manual/Manual_fasma.tex
@@ -4,6 +4,7 @@
 \usepackage[utf8]{inputenc}
 \usepackage{listings}
 \usepackage{color}
+\usepackage{hyperref}
 \definecolor{backcolour}{rgb}{0.95,0.95,0.92}
 \definecolor{codegreen}{rgb}{0,0.6,0}
 \lstdefinestyle{mystyle}{
@@ -27,61 +28,90 @@
 
 \section{The configuration file}
 
-To use FASMA in an automatic way, configure the \texttt{StarMe\_synth.cfg}. This file should be in the folder where FASMA runs. There are various configurations depending on what the user wants. 
-A mandatory requirement is the line list. 
+To use FASMA, configure the \texttt{config.cfg} which is in YAML format. Because this format is whitespace sensitive, you can check if the format is correct \href{http://www.yamllint.com/}{here}.
+This file must be in the folder where FASMA runs. There are various configuration combinations depending on what the user wants. A mandatory requirement is the line list and the interval 
+file, where the whole path needs to be inserted. 
 
 \begin{itemize}
- \item To create a synthetic spectrum with solar values, only the line list file is required, e.g.:
+ \item This is the simplest option for FASMA where a synthetic spectrum is created. The rest options are set to default.
+ {\footnotesize
  \begin{lstlisting}[language=Python]
- #linelist
-
- linelist_sun_arcturus_calib.lst
+#
+star:
+  linelist: /home/FASMA_synthesis/FASMA/rawLinelist/linelist.lst
+  intervals_file: /home/FASMA_synthesis/FASMA/rawLinelist/intervals.lst
 \end{lstlisting}
-The complete path should be given for the list file in the configuration file (\texttt{StarMe\_synth.cfg}). The default line list is in the \texttt{rawLinelist} folder already provided within FASMA.
- \item For a given line list select your options. The options are space seperated from the line list file and the options are comma seperated, e.g.:
+}
+
+The default line list is in the \texttt{rawLinelist} folder already provided within FASMA.
+ \item For a given line list select your options, e.g.:
  {\footnotesize
  \begin{lstlisting}[language=Python]
  #linelist options
-
- linelist_sun_arcturus_calib.lst observations:Sun_HARPS.fits,resolution:115000,
- minimize,refine,inter_file:intervals.lst,plot
+star:
+  linelist: /home/FASMA_synthesis/FASMA/rawLinelist/linelist.lst
+  intervals_file: /home/FASMA_synthesis/FASMA/rawLinelist/intervals.lst
+  fix_vmac: true
+  fix_vt: true
+  minimize: true
+  refine: true
+  observations: /home/FASMA_synthesis/FASMA/spectra/Sun_HARPS.fits
+  resolution: 115000
+  plot: true
 \end{lstlisting}}
-This example indicates that for the observed spectrum \texttt{Sun\_HARPS.fits} with resolution 115000, FASMA will minimize it starting from solar values using the refine option (see below). The 
-spectral regions for the synthesis are difined in the intervals.lst file with the option inter\_file. The final spectrum will be plotted. 
-For the not listed options, FASMA uses the default ones. {\bf The complete paths of the line list, spectrum, and intervals file should be given.} FASMA already includes the 
-line list (\texttt{linelist\_sun\_arcturus\_calib.lst}), the interval file (\texttt{intervals.lst}) and a HARPS spectrum of the Sun (\texttt{Sun\_HARPS.fits})
 
- \item For a given line list and a set of initial values, create a synthetic spectrum. The initial parameters are space seperated from the line list file, e.g.:
- \begin{lstlisting}[language=Python]
[…]
