+++ FASMA/interpolation.py
@@ -338,43 +338,3 @@ def save_model(model, params, abund=0.0, elem=False, type='apogee_kurucz', fout=
     np.savetxt(
         fout, model, header=header, footer=footer, comments='', delimiter=' ', fmt=_fmt
     )
-
-
-if __name__ == '__main__':
-    import argparse
-
-    args = argparse.ArgumentParser(description='Get a model atmosphere.')
-    args.add_argument('teff', type=int, help='Effective temperature')
-    args.add_argument('logg', type=float, help='Surface gravity')
-    args.add_argument('feh', type=float, help='Metallicity, [Fe/H]')
-    args.add_argument('vt', type=float, help='Microturbulence')
-    args.add_argument('-elem', type=str, help='Element', default=False)
-    args.add_argument('-abund', type=float, help='Abundance', default=0.0)
-    args.add_argument('-o', '--out', help='Output atmosphere', default='out.atm')
-    args.add_argument(
-        '-a',
-        '--atmosphere',
-        help='Model atmosphere',
-        choices=['apogee_kurucz', 'marcs'],
-        default='apogee_kurucz',
-    )
-    args = args.parse_args()
-
-    params = [args.teff, args.logg, args.feh, args.vt]
-    atmosphere, p = interpolator(
-        params,
-        elem=args.elem,
-        abund=args.abund,
-        save=False,
-        atmtype=args.atmosphere,
-        result=True,
-    )
-    save_model(
-        atmosphere,
-        params,
-        elem=args.elem,
-        abund=args.abund,
-        type=args.atmosphere,
-        fout=args.out,
-    )
-    print('Atmosphere model sucessfully saved in: %s', args.out)
