### 2020-02-13 #2048 @MariaTsantaki
Dear reviewers,
Thank you very much for the comments. They helped to improve the code. The major changes were on the setup, how FASMA function works in scripts, and the configuration file. This version seems easier to work with.

1) The configuration file (renamed from StarMe.cfg to config.yml) has now a yaml format with all the keys/options available. The file contains the default options.

2) There were some classes renamed and as suggested, now FASMA can accept options and update the default ones.
e.g.
```
from FASMA import FASMA
options = {'teff':4500, 'fix_teff':True}
result = FASMA(**options)
```

3) Setting up MOOG is tricky, hopefully now it will work better. The shell script install_fasma.sh will fill the moog path but will ask for the system type. There is a hidden limit for the length of the path characters though for fortran. I commented out the places where sm was used. I checked installed FASMA for versions with and without sudo: pip install . and pip install --user .
They both work but I leave it as the first for now.

There were indeed some dependencies which were not used and now are removed. Unfortunately, I could not remove for instance astropy as it manipulates fits files and I am not familiar with other methods dealing with them.

A good test (once installation runs fine) will be to adapt the config.yml with the correct paths (for linelist, intervals_file, observations) and change the option 'minimize' to True. Then by running fasma the parameters of the Sun will be obtained!

Thank you very much for your time!
