### 2020-02-14 MariaTsantaki/FASMA-synthesis#12 @warrickball
I just built FASMA and the example code ran to completion so the recent changes appear to have fixed this issue.
