+++ FASMA/MOOG/Begin.f
@@ -68,8 +68,7 @@ c*****clear the text screen
 
 c*****open data files carried with the source code: Barklem damping
       nfbarklem = 35
-      num = 60
-      call getcount (num,moogpath)
+      num=len(trim(moogpath))
       if (moogpath(num:num) .ne. '/') then
          num = num + 1
          moogpath(num:num) = '/'
@@ -81,8 +80,7 @@ c*****open data files carried with the source code: Barklem damping
 
 c*****open data files carried with the source code: Barklem UV damping
       nfbarklemUV = 36
-      num = 60
-      call getcount (num,moogpath)
+      num=len(trim(moogpath))
       if (moogpath(num:num) .ne. '/') then
          num = num + 1
          moogpath(num:num) = '/'
