### 2020-02-18 #2048 @ipelupessy
I managed to install and run with the provided instructions. The path limit is annoying; I have made a PR (https://github.com/MariaTsantaki/FASMA-synthesis/pull/17) to fix this..it adds another change in the MOOG source though (maybe its good to communicate it upstream to MOOG developer)

I can confirm @warrickball problem; I think the fits file does have a WAVELMIN keyword in the header with the smae info as CRVAL1 ??
