### 2020-02-18 #2048 @ipelupessy
one comment about the code paper proof: it doesn't mention what the improvement/changes are from the previous published version. I think this is needed.
