### 2020-02-18 #2048 @ipelupessy
The following issues are open from my checklist:

Functionality.Functionality: related to documentation below I am not sure what all the functionality of the packages is;

Documentation.A statement of need: Its not clearly stated what the problem the software is designed to solve, e.g.: "FASMA delivers the atmospheric stellar parameters (effective temperature, surface gravity, metallicity, microturbulence, macroturbulence, rotational velocity) and chemical abundances of 12 elements." -> from what? Reader has to guess. What is the fundamental method of the package (multidimensional fit?)

Documentation.Functionality documentation: The manual provides a description of the configuration file. missing is a high level description of what the config.yml is needed for (e.g.: "The configuration file provides the input for the fitting procedure and..?")

Documentation.Automated tests: please provide if there are any tests beyond the basic example. (if not its ok)

Documentation.Community guidelines: not clear how to contribute

Software paper.Summary: atm is geared to specialist. Please provide one or two lines describing the software for non-specialist

Software paper.A statement of need: both the problem and the audience can be clarified a bit more

These are not huge issues, but I think they should be adressed to make the functionality of the software a bit more clear.

This concludes my review.
