### 2020-02-18 #2048 @ipelupessy
@xuanxu: I have my final comments out above; if these are adressed I think the checklist is ok. Authors have made very good progress on improving the installation. I think the minor points mentioned above will also be addressed promptly ;-)
