### 2020-02-18 #2048 @warrickball
I managed to get FASMA to run renaming the file to `.spec` and changing these lines in `read_observations` (`observations.py`):

````
            flux = x['FLUX'][0] # x['flux']
            wave = x['WAVE'][0] # x['wavelength']
````
The run eventually crashes with

````
Iter       4    CHI-SQUARE =  77689.37109  DOF =  26425
   Teff = 5780.200516
   logg = 3.08652508
   [Fe/H] = 0.1202310063
   vt = 0
   vmac = 11.79304433
   vsini = 1.201662041
    Teff:  5880.2   logg: 3.09   [Fe/H]: 0.12   vt: 0.00   vmac: 11.79   vsini: 1.20
    Teff:  5680.2   logg: 3.09   [Fe/H]: 0.12   vt: 0.00   vmac: 11.79   vsini: 1.20
    Teff:  5780.2   logg: 3.24   [Fe/H]: 0.12   vt: 0.00   vmac: 11.79   vsini: 1.20
    Teff:  5780.2   logg: 2.94   [Fe/H]: 0.12   vt: 0.00   vmac: 11.79   vsini: 1.20
    Teff:  5780.2   logg: 3.09   [Fe/H]: 0.22   vt: 0.00   vmac: 11.79   vsini: 1.20
    Teff:  5780.2   logg: 3.09   [Fe/H]: 0.02   vt: 0.00   vmac: 11.79   vsini: 1.20
    Teff:  5780.2   logg: 3.09   [Fe/H]: 0.12   vt: 0.20   vmac: 11.79   vsini: 1.20
    Teff:  5780.2   logg: 3.09   [Fe/H]: 0.12   vt: 0.00   vmac: 11.79   vsini: 1.20
    Teff:  5780.2   logg: 3.09   [Fe/H]: 0.12   vt: 0.00   vmac: 12.29   vsini: 1.20
    Teff:  5780.2   logg: 3.09   [Fe/H]: 0.12   vt: 0.00   vmac: 11.29   vsini: 1.20
    Teff:  5780.2   logg: 3.09   [Fe/H]: 0.12   vt: 0.00   vmac: 11.79   vsini: 2.70
    Teff:  5780.2   logg: 3.09   [Fe/H]: 0.12   vt: 0.00   vmac: 11.79   vsini: 0.00
    Teff:  5769.7   logg: 2.34   [Fe/H]: 0.08   vt: 0.00   vmac: 20.00   vsini: 1.02
Traceback (most recent call last):
  File "/home/wball/.local/bin/fasma", line 12, in <module>
    driver = FASMA.FASMA(cfgfile=cfgfile, overwrite=None)
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthDriver.py", line 49, in __init__
    self.synthdriver()
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthDriver.py", line 575, in synthdriver
    self.status = self.minizationRunner()
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthDriver.py", line 383, in minizationRunner
    self.params, self.xo, self.yo = function.minimize()
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/minimization.py", line 349, in minimize
    maxiter=20,
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/mpfit.py", line 1306, in __init__
    [self.status, wa4] = self.call(fcn, self.params, functkw)
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/mpfit.py", line 1578, in call
    [status, f] = fcn(x, fjac=fjac, **functkw)
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/minimization.py", line 251, in myfunct
    p, atmtype=self.model, driver='synth', ranges=self.ranges, **options
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/utils.py", line 570, in fun_moog_synth
    epsilon=options['limb'],
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthetic.py", line 259, in broadening
    y_mac = vmac_broadening(x, y, vmacro_rad=vmac)
  File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthetic.py", line 240, in vmac_broadening
    ("Spectrum range too narrow for macroturbulent broadening")
ValueError: Spectrum range too narrow for macroturbulent broadening
````
but I'm not a spectroscopist, so I'm not sure whether this is my fault. The initial parameters I used are a rough average of the values under [HD 38529's SIMBAD entry](http://simbad.u-strasbg.fr/simbad/sim-id?Ident=%40826786&Name=HD++38529&submit=display+all+measurements#lab_meas) but the fit appears to diverging from these values.
