+++ FASMA/observations.py
@@ -11,6 +11,132 @@ import matplotlib
 matplotlib.use('TkAgg')
 import matplotlib.pyplot as plt
 
+def eso_fits(hdulist):
+    '''A little demo utility to illustrate the ESO SDP 1D spectrum file format.
+    2017-Aug-08, archive(at)eso.org
+    The ESO 1D spectral format is compliant with the IVOA Spectrum data model.
+    Both the SPECTRUM 1.0 and SPECTRUM 2.0 versions are supported.
+    References:
+    ESO SDP standard: http://www.eso.org/sci/observing/phase3/p3sdpstd.pdf
+    ESO SDP FAQ page: http://www.eso.org/sci/observing/phase3/faq.html
+    This script:      http://archive.eso.org/cms/eso-data/help/1dspectra.html
+
+    Output
+    -----
+    wave : raw observed wavelength
+    flux : raw observed flux
+    '''
+
+    phu = hdulist[0].header           # Primary Header Unit: metadata
+    scihead = hdulist[1].header       # Header of the first FITS extension: metadata
+    scidata = hdulist[1].data         # Data in the first FITS extension: the spectrum
+
+    # Checking some compliance
+    # 1. keyword PRODCATG must be present in primary header unit
+    prodcatg = phu['PRODCATG']
+    # 2. value of keyword PRODCATG must match SCIENCE.SPECTRUM*
+    if not prodcatg.startswith('SCIENCE.SPECTRUM'):
+        errorMessage = "Expected header keyword: PRODCATG = 'SCIENCE.SPECTRUM'\nFound: PRODCATG = '%s'\nFile not compliant with the 1d spectrum specifications\nof the ESO Science Data Product standard." % prodcatg
+        print('filename = %s   NOT COMPLIANT' % filename)
+        print(errorMessage)
+        return None
+
+    # 3. Various keywords must be defined, among them the ones here below:
+    try:
+        origfile = phu['ORIGFILE']    # Original filename as assigned by the data provider
+        instrume = phu['INSTRUME']  # Name of the instrument
+        wavelmin = phu['WAVELMIN']  # Minimum wavelength in nm
+        wavelmax = phu['WAVELMAX']  # Maximum wavelength in nm
+        respower = phu['SPEC_RES']  # Spectral resolving power (lambda / delta_lambda)
+        snr      = phu['SNR']       # Signal to Noise Ratio
+        specaxisucd  = scihead['TUCD1'] # Gives the type of spectral axis (see SPECTRAL AXIS below)
+    except:
+        errorMessage='File not compliant with the 1D spectrum specifications of the ESO Science Data Product standard; some of the mandatory keywords were not found in primary header unit'
+        print('filename = %s   NOT COMPLIANT' % filename)
+        print('ERROR = %s' % (errorMessage))
+        return None
+
+    # SPECTRAL AXIS: could be either wavelength, frequency, or energy;
+    # if wavelength, the distinction between wavelength in air or in vacuum is provided by the presence of the obs.atmos token in the TUCD1.
+    # the variable spectype will carry to whole info.
+    spectype = None
+    if specaxisucd.startswith('em.wl'):
+        if specaxisucd == 'em.wl':
+            spectype = 'wavelength in vacuum (TUCD1=%s)' % specaxisucd
+        elif specaxisucd == 'em.wl;obs.atmos':
+            spectype = 'wavelength in air (TUCD1=%s)' % specaxisucd
+        else:
+            spectype = 'wavelength (TUCD1=%s)' % specaxisucd
+    elif specaxisucd.startswith('em.freq'):
+        spectype = 'frequency (TUCD1=%s)' % specaxisucd
[…]
