### 2020-02-19 #2048 @MariaTsantaki
> The test run of the Sun runs fine but I tried to run using a [HARPS spectrum of HD38529](https://dataportal.eso.org/dataPortal/file/ADP.2018-11-14T01:07:37.242) that I downloaded off the [ESO Archive](http://archive.eso.org/) and got the error
>
> ```
> /home/wball/.local/lib/python3.7/site-packages/pyfits/__init__.py:22: PyFITSDeprecationWarning: PyFITS is deprecated, please use astropy.io.fits
>   PyFITSDeprecationWarning)  # noqa
> Traceback (most recent call last):
>   File "/home/wball/.local/bin/fasma", line 12, in <module>
>     driver = FASMA.FASMA(cfgfile=cfgfile, overwrite=None)
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthDriver.py", line 49, in __init__
>     self.synthdriver()
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthDriver.py", line 530, in synthdriver
>     for (self.initial, self.options) in self._genStar():
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthDriver.py", line 124, in _genStar
>     self._setup(line)
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthDriver.py", line 104, in _setup
>     self._options(line)
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/synthDriver.py", line 245, in _options
>     defaults['snr'] = snr(defaults['observations'])
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/observations.py", line 286, in snr
>     snr = [sub_snr(snr_region) for snr_region in snr_regions]
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/observations.py", line 286, in <listcomp>
>     snr = [sub_snr(snr_region) for snr_region in snr_regions]
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/observations.py", line 247, in sub_snr
>     wave_cut, flux_cut, l = read_observations(fname, w1, w2)
>   File "/home/wball/.local/lib/python3.7/site-packages/FASMA/observations.py", line 122, in read_observations
>     start_wave = header['CRVAL1']  # initial wavelenght
>   File "/home/wball/.local/lib/python3.7/site-packages/astropy/io/fits/header.py", line 148, in __getitem__
>     card = self._cards[self._cardindex(key)]
>   File "/home/wball/.local/lib/python3.7/site-packages/astropy/io/fits/header.py", line 1708, in _cardindex
>     raise KeyError(f"Keyword {keyword!r} not found.")
> KeyError: "Keyword 'CRVAL1' not found."
> ```
>
> I presume that FASMA expects the spectrum to be in a particular format (that includes the `CRVAL1` key in the FITS header) but this isn't documented anywhere.
>

That's a very nice catch. It has been some time since ESO has been standardizing their reduced fits files but I haven't been aware of that because I have been using older data. I followed their recommendations to read such files. https://archive.eso.org/cms/eso-data/help/1dspectra.html
I push the changes for this.
