+++ paper.md
@@ -30,18 +30,21 @@ bibliography: paper.bib
 
 # Summary
 
-Effective temperature (Teff), surface gravity (logg), and metallicity ([M/H]) are basic stellar atmospheric parameters necessary to characterize a star. Once these parameters are obtained, we can in turn, infer chemical abundances of various elements. In this work, we use spectroscopy as a powerful tool to extract this information from stellar atmospheres for stars with spectral type FGK both dwarfs and giants.
+Effective temperature (Teff), surface gravity (logg), and metallicity ([M/H]) are basic stellar atmospheric parameters necessary to characterize a star. Once these parameters are obtained, we can in turn, infer their chemical abundances of various elements and in conjunction with evolutionary models to estimate their evolution, i.e. mass and radius. In this work, we use spectroscopy as a powerful tool to extract this information from stellar atmospheres applied to stars with spectral type FGK both dwarfs and giants. The growing number of spectroscopic surveys dedicated to the study of the Galactic stellar populations has inflated the number of high quality spectra to several hundreds of thousands. This amount is expected to multiply with the forthcoming surveys, such as WEAVE [@deJong:2019] and 4MOST [@Dalton:2014]. The success of these surveys highly depends on the analysis tools to exploit sufficiently all spectral information. Moreover, it is a well-known axiom in exoplanetary studies that one can only determine the planetary properties once the ones of the host star are known. The planetary properties such as mass, radius, composition, are directly linked to their hosts and therefore, robust tools for the derivation of these parameters are necessary.
 
-``FASMA`` is a Python package to derive stellar atmospheric parameters based on the spectral synthesis technique. The principle of this technique relies on the comparison between synthetic spectra with observations to yield the best-fit parameters under a minimization process.
+Our goal is to provide a tool to determine accurately and precisely the stellar parameters and chemical abundances to fulfill the above purposes. ``FASMA`` is a Python package to derive the main stellar atmospheric parameters based on the spectral synthesis technique. The principle of this technique relies on the comparison of synthetic spectra with observations to yield the best-fit parameters under a &chi;<sup>2</sup> minimization process:
 
-The synthetic spectra are created on-the-fly with a radiative transfer code, [``MOOG``](https://www.as.utexas.edu/~chris/moog.html) [@Sneden:1973] in fortran for a set of stellar parameters. The synthetic spectra are later convolved with different rotational profiles, such as macroturbulence (vmac), projected rotational velocity (vsini), and instrumental broadening (resolution) to match the observations. Spectral synthesis has model dependence and thus, grids of pre-computed stellar atmospheres, such as Kurucz [@Kurucz:1993] or marcs [@Gustafsson:2008] models, are provided with ``FASMA`` along this the interpolation processes for the models (stellar atmospheres) in the grid.
+ &chi;<sup>2</sup> = &Sigma; (obs<sub>i</sub> - synth<sub>i</sub>)<sup>2</sup>/&sigma;<sub>i</sub>
+
+where obs is the observed spectrum, synth the synthetic, and &sigma; the error on the observed flux for each spectral point, i. The best parameters are the ones that minimize the non-linear least-squares based on the Levenberg-Marquardt algorithm [@Marquardt:1963].
+
[…]
