+++ paper.bib
@@ -1,3 +1,144 @@
+@ARTICLE{deJong:2019,
+       author = {{de Jong}, R.~S. and {Agertz}, O. and {Berbel}, A.~A. and {Aird}, J. and
+         {Alexander}, D.~A. and {Amarsi}, A. and {Anders}, F. and {Andrae}, R. and
+         {Ansarinejad}, B. and {Ansorge}, W. and {Antilogus}, P. and {Anwand
+        -Heerwart}, H. and {Arentsen}, A. and {Arnadottir}, A. and
+         {Asplund}, M. and {Auger}, M. and {Azais}, N. and {Baade}, D. and
+         {Baker}, G. and {Baker}, S. and {Balbinot}, E. and {Baldry}, I.~K. and
+         {Banerji}, M. and {Barden}, S. and {Barklem}, P. and
+         {Barth{\'e}l{\'e}my-Mazot}, E. and {Battistini}, C. and {Bauer}, S. and
+         {Bell}, C.~P.~M. and {Bellido-Tirado}, O. and {Bellstedt}, S. and
+         {Belokurov}, V. and {Bensby}, T. and {Bergemann}, M. and
+         {Bestenlehner}, J.~M. and {Bielby}, R. and {Bilicki}, M. and
+         {Blake}, C. and {Bland-Hawthorn}, J. and {Boeche}, C. and {Boland}, W. and
+         {Boller}, T. and {Bongard}, S. and {Bongiorno}, A. and {Bonifacio}, P. and
+         {Boudon}, D. and {Brooks}, D. and {Brown}, M.~J.~I. and {Brown}, R. and
+         {Br{\"u}ggen}, M. and {Brynnel}, J. and {Brzeski}, J. and
+         {Buchert}, T. and {Buschkamp}, P. and {Caffau}, E. and {Caillier}, P. and
+         {Carrick}, J. and {Casagrande}, L. and {Case}, S. and {Casey}, A. and
+         {Cesarini}, I. and {Cescutti}, G. and {Chapuis}, D. and
+         {Chiappini}, C. and {Childress}, M. and {Christlieb}, N. and
+         {Church}, R. and {Cioni}, M. -R.~L. and {Cluver}, M. and {Colless}, M. and
+         {Collett}, T. and {Comparat}, J. and {Cooper}, A. and {Couch}, W. and
+         {Courbin}, F. and {Croom}, S. and {Croton}, D. and {Daguis{\'e}}, E. and
+         {Dalton}, G. and {Davies}, L.~J.~M. and {Davis}, T. and
+         {de Laverny}, P. and {Deason}, A. and {Dionies}, F. and {Disseau}, K. and
+         {Doel}, P. and {D{\"o}scher}, D. and {Driver}, S.~P. and {Dwelly}, T. and
+         {Eckert}, D. and {Edge}, A. and {Edvardsson}, B. and
+         {Youssoufi}, D.~E. and {Elhaddad}, A. and {Enke}, H. and
+         {Erfanianfar}, G. and {Farrell}, T. and {Fechner}, T. and {Feiz}, C. and
+         {Feltzing}, S. and {Ferreras}, I. and {Feuerstein}, D. and
+         {Feuillet}, D. and {Finoguenov}, A. and {Ford}, D. and
+         {Fotopoulou}, S. and {Fouesneau}, M. and {Frenk}, C. and {Frey}, S. and
+         {Gaessler}, W. and {Geier}, S. and {Fusillo}, N.~G. and {Gerhard}, O. and
+         {Giannantonio}, T. and {Giannone}, D. and {Gibson}, B. and
+         {Gillingham}, P. and {Gonz{\'a}lez-Fern{\'a}ndez}, C. and
[…]
