### 2020-02-19 #2048 @MariaTsantaki
@ipelupessy
Thanks again for the comments.
I added more information on the Summary for the purpose of this tool and more details for the technical part. I also updated the README and manual.
I am not sure how other tests would be useful but I can include more spectra to test more than one star.
