### 2020-02-19 #2048 @MariaTsantaki
@warrickball You almost made it! FASMA reached its limit for macroturbulence and it broke. There were not efficient number spectral points for the calculations. Since 20 km/s is too high for 'normal stars' I suspect there is another problem here. The spectrum region can be too narrow if other velocities with similar broadening profiles broaden the lines too much such as rotation. In this case I would recommend to increase the ranges in the intervals.lst file. However, in this case I suspect another problem. I neglected to mention that the input spectra for FASMA have to be corrected for radial velocities. You can see if it is corrected or not by just plotting the spectrum with FASMA  (configure the config.yml file) and see that it is shifted compared to the synthetic one.
