### 2020-02-20 #2048 @warrickball
I expect that the file format issues would restrict FASMA to some specific subset of spectra or require the user to modify their data for the command line script to run. It would be better to specify some FASMA data format to separate that from the analysis and define that format somewhere in the documentation. e.g. if I know that FASMA can use a plain text file with columns of wavelength (in some units) versus flux (in some units), then I can deal with converting my spectra to that format and leave FASMA to analyse it. This basically already exists but isn't documented. I looked at the source code to infer the appropriate FITS structure.
