### 2020-02-20 #2048 @ipelupessy
..and/or accept a dictionary with the needed data as input for 'observations' ..
