### 2020-02-20 #2048 @warrickball
After I shifted the wavelength values in the spectrum by 30 km/s, I got a decent fit to my star, so I think I'm now happy that FASMA is capable of fitting the fundamental parameters of stellar spectra. This did, however, involve writing myself some scripts to convert between various formats and apply the RV correction. Assuming the RV correction is as simple as multiplying by (1+v/c), it would be nice if FASMA could do this for the user.

I'm now testing whether FASMA can determine individual abundances, as said in the paper. I worked out (undocumented) that I needed to change `intervals_file` from `intervals.lst` to `intervals_elements.lst` and `linelist` from `linelist.lst` to `elements.lst`. I got a result for `Na`, `Mg`, `Sc`, `Ti`, `V`, `Cr` and `Mn` but for `Al`, `Si`, `Ca` and `Ni` I get errors like this for the Sun:
````
Line list: /home/wball/work/FASMA-synthesis/FASMA/rawLinelist/elements.lst
Intervals list: /home/wball/work/FASMA-synthesis/FASMA/rawLinelist/intervals_elements.lst
Linelist contains 70 lines in 2 intervals
SNR: 261
Observed spectrum contains 800 points
Starting minimization for element abundance...
    [Al/H]: 0.00
Traceback (most recent call last):
  File "/home/wball/.local/bin/fasma", line 7, in <module>
    exec(compile(f.read(), __file__, 'exec'))
  File "/home/wball/work/FASMA-synthesis/bin/fasma", line 12, in <module>
    driver = FASMA.FASMA(cfgfile=cfgfile, overwrite=None)
  File "/home/wball/work/FASMA-synthesis/FASMA/synthDriver.py", line 49, in __init__
    self.synthdriver()
  File "/home/wball/work/FASMA-synthesis/FASMA/synthDriver.py", line 601, in synthdriver
    self.status = self.minizationElementRunner()
  File "/home/wball/work/FASMA-synthesis/FASMA/synthDriver.py", line 436, in minizationElementRunner
    self.elemabund, self.xo, self.yo = function.minimizeElement()
  File "/home/wball/work/FASMA-synthesis/FASMA/minimization.py", line 389, in minimizeElement
    maxiter=20,
  File "/home/wball/work/FASMA-synthesis/FASMA/mpfit.py", line 1031, in __init__
    [self.status, fvec] = self.call(fcn, self.params, functkw)
  File "/home/wball/work/FASMA-synthesis/FASMA/mpfit.py", line 1578, in call
    [status, f] = fcn(x, fjac=fjac, **functkw)
  File "/home/wball/work/FASMA-synthesis/FASMA/minimization.py", line 254, in myfunct
    sl = InterpolatedUnivariateSpline(xs, ys, k=1)
  File "/home/wball/.local/lib/python3.7/site-packages/scipy/interpolate/fitpack2.py", line 601, in __init__
    raise ValueError('x must be strictly increasing')
ValueError: x must be strictly increasing
````
Here's my configuration file, mostly adapted from the example:
````yaml
#
star1:
    teff: 5777
    logg: 4.44
    feh: 0.0
    vt: 1.0
    vmac: 3.21
    vsini: 1.90
    model: apogee_kurucz
    MOOGv: 2014
    save: False
    # Na, Mg, Al, Si, Ca, Sc, Ti, V, Cr, Mn, Ni
    element: 'Al'
    fix_teff: False
    fix_logg: False
    fix_feh: False
    fix_vt: False
    fix_vmac: False
    fix_vsini: False
    plot: True
    plot_res: False
    damping: 1
    step_wave: 0.01
    step_flux: 3.0
    minimize: False
    refine: False
    observations: /home/wball/work/FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits
    intervals_file: /home/wball/work/FASMA-synthesis/FASMA/rawLinelist/intervals_elements.lst
    linelist: /home/wball/work/FASMA-synthesis/FASMA/rawLinelist/elements.lst
    snr: null
    resolution: null
    limb: 0.6
````
