### 2020-02-21 #2048 @warrickball
Now that I could finally install and confirm that the code runs and is reasonably usable, here's my first pass at reviewing it from beginning to end and identifying some outstanding issues. I've ticked off things I'm already satisfied with and commented below on improvements that could be made (including things that are already satisfactory but could be better).

> License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

I'm not a licensing expert but I'm not really sure how to accommodate MOOG's unclear licensing status. I think the best solution would be that the FASMA repo doesn't contain MOOG but FASMA also needs to modify MOOG to remove the SuperMongo dependency. I'm not sure what to suggest but I remain uncomfortable until an expert has weighed in.

> Installation: Does installation proceed as outlined in the documentation?

Mostly.  I ran the `pip install` command myself after installation so that I could add the `--user` flag.  It would be nice if the install script would somehow allow arguments like this to be passed and even
better if the whole installation could be handled by `pip`. Or just remove the `pip` command from the install script and change the documentation to say "run the install script, then `pip install` FASMA".

> Functionality: Have the functional claims of the software been confirmed?

Not quite. I get errors when I try to determine the individual abundances of `Al`, `Si`, `Ca` and `Ni` in the bundled solar spectrum.

> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

While there aren't any performance claims, there might be places where the code could be sped up. For example, it looks like the observed spectrum is re-read many times while the code is running.  This isn't noticeable for FITS files, which are binary, but it was noticeable when reading from a ~115,000 line plain text HARPS spectrum.

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

The requirements for FASMA's Python package are clear (and I appreciate that there are relatively few). There is no clear list of what MOOG requires but I imagine any but the most ancient of Fortran compilers will do.

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The authors include one example of fitting the fundamental properties but not of fitting individual abundances, which for me doesn't currently work for all elements.

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

At the moment, I find the documentation unsatisfactory.  Though there is a PDF manual that lists the options for the configuration file and a top-level ``README.md`` with some other details, neither specifies some important details like expected file formats.  The ``README.md`` has a few lines about using the Python functions directly but I couldn't find any documentation for the Python API.  On the other hand, the authors don't motivate using the Python interface instead of the command line program so I'd be satisfied if they just dropped the Python interface example.  Otherwise, I expect API documentation somewhere.

I also find it slightly frustrating having to flip back and forth between two distinct sources of documentation: the ``README.md`` and the PDF "manual". Ideally there'd be Sphinx documentation rendered somewhere.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

There don't appear to be any automated tests of the main functionality but the fact that the example runs to completion suffices. If a similar example is created for the individual solar abundances, I would be satisfied.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

No.

> Software paper
> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I think the value of spectroscopic analysis can be made more clearly, even though the authors do already say it's one of the basic things we rely upon in astronomy. Example subfields in which I can think of spectroscopy having a big influence are in constraining stellar physics and distinguishing stellar populations.  Detecting companions (be they stars or planets) also requires spectroscopic analysis but with the radial velocity, which FASMA doesn't handle.

I think it's also worth pointing out that spectroscopic analysis is currently not very accessible.  So, though codes like MOOG exist, they are relatively difficult to use.  The real value in a code like FASMA is to make the process that much easier.  Even though I've never used MOOG directly and FASMA's documentation is only partially complete, it still enabled me to fit a stellar spectrum, which is an achievement!

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

The authors don't compare FASMA to other software.  I'm not a specialist myself but perhaps [Spectroscopy Made Easy](http://www.stsci.edu/~valenti/sme.html) is similar?  Or [CMFGEN](http://kookaburra.phyast.pitt.edu/hillier/web/CMFGEN.htm) (though that might be for hot stars)?

> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

I would say the paper would benefit from language editing. (To be fair, most papers—including my own—would.)

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

There's no reference for the implementation of the Levenberg–Marquardt algorithm (currently in [`mpfit.py`](https://github.com/MariaTsantaki/FASMA-synthesis/blob/841ce323b63d62cd7310ae64f5865ebadcb67331/FASMA/mpfit.py). I presumed the authors would use a standard Python package for this (e.g. [`scipy.optimize.leastsq`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.leastsq.html) or [`lmfit`](https://lmfit.github.io/lmfit-py/)) but this implementation comes from some other (not clearly licensed) source.
