### 2020-02-24 #2048 @MariaTsantaki
> After I shifted the wavelength values in the spectrum by 30 km/s, I got a decent fit to my star, so I think I'm now happy that FASMA is capable of fitting the fundamental parameters of stellar spectra. This did, however, involve writing myself some scripts to convert between various formats and apply the RV correction. Assuming the RV correction is as simple as multiplying by (1+v/c), it would be nice if FASMA could do this for the user.
>
> I'm now testing whether FASMA can determine individual abundances, as said in the paper. I worked out (undocumented) that I needed to change `intervals_file` from `intervals.lst` to `intervals_elements.lst` and `linelist` from `linelist.lst` to `elements.lst`. I got a result for `Na`, `Mg`, `Sc`, `Ti`, `V`, `Cr` and `Mn` but for `Al`, `Si`, `Ca` and `Ni` I get errors like this for the Sun:
>
> ```
> Line list: /home/wball/work/FASMA-synthesis/FASMA/rawLinelist/elements.lst
> Intervals list: /home/wball/work/FASMA-synthesis/FASMA/rawLinelist/intervals_elements.lst
> Linelist contains 70 lines in 2 intervals
> SNR: 261
> Observed spectrum contains 800 points
> Starting minimization for element abundance...
>     [Al/H]: 0.00
> Traceback (most recent call last):
>   File "/home/wball/.local/bin/fasma", line 7, in <module>
>     exec(compile(f.read(), __file__, 'exec'))
>   File "/home/wball/work/FASMA-synthesis/bin/fasma", line 12, in <module>
>     driver = FASMA.FASMA(cfgfile=cfgfile, overwrite=None)
>   File "/home/wball/work/FASMA-synthesis/FASMA/synthDriver.py", line 49, in __init__
>     self.synthdriver()
>   File "/home/wball/work/FASMA-synthesis/FASMA/synthDriver.py", line 601, in synthdriver
>     self.status = self.minizationElementRunner()
>   File "/home/wball/work/FASMA-synthesis/FASMA/synthDriver.py", line 436, in minizationElementRunner
>     self.elemabund, self.xo, self.yo = function.minimizeElement()
>   File "/home/wball/work/FASMA-synthesis/FASMA/minimization.py", line 389, in minimizeElement
>     maxiter=20,
>   File "/home/wball/work/FASMA-synthesis/FASMA/mpfit.py", line 1031, in __init__
>     [self.status, fvec] = self.call(fcn, self.params, functkw)
>   File "/home/wball/work/FASMA-synthesis/FASMA/mpfit.py", line 1578, in call
>     [status, f] = fcn(x, fjac=fjac, **functkw)
>   File "/home/wball/work/FASMA-synthesis/FASMA/minimization.py", line 254, in myfunct
>     sl = InterpolatedUnivariateSpline(xs, ys, k=1)
>   File "/home/wball/.local/lib/python3.7/site-packages/scipy/interpolate/fitpack2.py", line 601, in __init__
>     raise ValueError('x must be strictly increasing')
> ValueError: x must be strictly increasing
> ```
You are right. There were overlapping wavelength intervals for the synthesis and the interpolation was crashing. I fixed this in the ```synthetic.py``` file when reading the intervals now they are merged when there are overlapping ones. Thanks! I also added from VALD some missing lines I noticed from the plots. And I checked now it works properly for the elements.

```
from FASMA import FASMA

 elements = ['Na', 'Mg', 'Al', 'Si', 'Ca', 'Sc', 'Ti', 'V', 'Cr', 'Mn', 'Ni']

 for el in elements:

     options = {'plot':True, 'linelist':'FASMA-synthesis/FASMA/rawLinelist/elements.lst', 'intervals_file':'FASMA-synthesis/FASMA/rawLinelist/intervals_elements.lst', 'element':el, 'resolution':115000, 'observations': 'FASMA-synthesis/FASMA/spectra/Sun_HARPS.fits'}

        r = FASMA(**options)
```
