### 2020-02-24 #2048 @MariaTsantaki
> Now that I could finally install and confirm that the code runs and is reasonably usable, here's my first pass at reviewing it from beginning to end and identifying some outstanding issues. I've ticked off things I'm already satisfied with and commented below on improvements that could be made (including things that are already satisfactory but could be better).
>
> > License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
>
> I'm not a licensing expert but I'm not really sure how to accommodate MOOG's unclear licensing status. I think the best solution would be that the FASMA repo doesn't contain MOOG but FASMA also needs to modify MOOG to remove the SuperMongo dependency. I'm not sure what to suggest but I remain uncomfortable until an expert has weighed in.
>
I am not sure what to do here too. I prefer to keep it inside FASMA because MOOG versions may change and also the SuperMongo dependency is now removed. I added some information about the MOOG code in its folder but if this is not enough, please let me know on how to proceed.
