### 2020-02-24 #2048 @MariaTsantaki
> > Functionality: Have the functional claims of the software been confirmed?
>
> Not quite. I get errors when I try to determine the individual abundances of `Al`, `Si`, `Ca` and `Ni` in the bundled solar spectrum.
>
True. We solved this with a recent commit.
