+++ README.md
@@ -4,68 +4,133 @@
 # FASMA 2.0
 Stellar spectral analysis package.
 
-FASMA delivers the atmospheric stellar parameters (effective temperature, surface gravity, metallicity, microturbulence, macroturbulence, and rotational velocity) based on the spectral synthesis technique. The principle of this technique relies on the comparison of synthetic spectra with observations to yield the best-fit parameters under a &chi;<sup>2</sup> minimization process. FASMA 2.0 is now updated to deliver also chemical abundances of 12 elements (Na, Mg, Al, Si, Ca, Sc, Ti, V, Cr, Mn, and Ni).
+FASMA delivers the atmospheric stellar parameters (effective temperature, surface gravity, metallicity, microturbulence, macroturbulence, and rotational velocity) based on the spectral synthesis technique. The principle of this technique relies on the comparison of synthetic spectra with observations to yield the best-fit parameters under a &chi;<sup>2</sup> minimization process. FASMA 2.0 is now updated to deliver also chemical abundances of 13 elements (Li, Na, Mg, Al, Si, Ca, Sc, Ti, V, Cr, Mn, and Ni).
 
 
 The python code runs with python 3 and is wrapped around the spectral synthesis package in fortran: [MOOG version 2019](http://www.as.utexas.edu/~chris/moog.html) (Sneden et al. 1973).
 
-The spectral synthesis technique requires model atmospheres for the calculation of the synthetic spectra by MOOG. FASMA includes two grids of models: Kurucz and marcs.
+The spectral synthesis technique requires model atmospheres for the calculation of the synthetic spectra by MOOG which is provided with FASMA. FASMA includes two grids of models: Kurucz and marcs.
+The instructions below are also included [here](https://github.com/MariaTsantaki/FASMA-synthesis/blob/master/manual/Manual_fasma.pdf).
 
 Contact here for bugs: tsantaki@arcetri.astro.it
 
 # Installation
-Installing FASMA requires a few simple steps. MOOG will also be installed to create the synthetic spectra.
+Installing FASMA requires 2 simple steps.
 
-Run the installation script `./install_fasma.sh`. The user will be asked about system requirements:
+1) Run the installation script `install_fasma.sh`, in the FASMA-synthesis folder, to unzip the model atmospheres and install MOOG (MOOG uses gfortan, f77, or g77 compilers). The user will be asked about system requirements:
 "rh64" for 64-bit linux system, "rh" for 32-bit linux system, "maclap" for mac laptop, "macdesk" for mac desktop.
+
+```sh
+$ ./install_fasma.sh
+```
+
+2) Then install FASMA with pip:
+```sh
+$ pip install .
+```
+
+Or, if you do not have sudo access, replace the pip command with:
+```sh
+$ pip install --user .
+```
+
 
 A list of basic python packages will be installed automatically with pip (see [requirements](https://github.com/MariaTsantaki/FASMA-synthesis/blob/master/requirements.txt)).
 
-More information about MOOG is [here](http://www.as.utexas.edu/~chris/moog.html).
-
-If you do not have sudo access, replace the pip command with:
-
-```
-pip install --user .
-```
-
 To uninstall FASMA:
 
-```
-pip uninstall FASMA
+```sh
+$ pip uninstall FASMA
 ```
 
 # Usage
 FASMA is so easy. You can run FASMA from the terminal by configuring the options in the `config.yml` file and then run in the working directory:
 
-```
-fasma
+```sh
+$ fasma
 ```
 
-A small tutorial is given [here](https://github.com/MariaTsantaki/FASMA-synthesis/blob/master/manual/Manual_fasma.pdf)
-FASMA includes a log file `fasma.log` every time it is used which catches errors in order to inform the user.
[…]
