+++ paper.bib
@@ -221,6 +221,27 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Delgado:2015,
+       author = {{Delgado Mena}, E. and {Bertr{\'a}n de Lis}, S. and {Adibekyan}, V. Zh. and
+         {Sousa}, S.~G. and {Figueira}, P. and {Mortier}, A. and
+         {Gonz{\'a}lez Hern{\'a}ndez}, J.~I. and {Tsantaki}, M. and
+         {Israelian}, G. and {Santos}, N.~C.},
+        title = "{Li abundances in F stars: planets, rotation, and Galactic evolution}",
+      journal = {\aap},
+     keywords = {stars: abundances, stars: fundamental parameters, stars: rotation, stars: evolution, planets and satellites: formation, planetary systems, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2015,
+        month = apr,
+       volume = {576},
+          eid = {A69},
+        pages = {A69},
+          doi = {10.1051/0004-6361/201425433},
+archivePrefix = {arXiv},
+       eprint = {1412.4618},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015A&A...576A..69D},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Marquardt:1963,
    author = {{Marquardt}, D.~W.},
     title = "{An Algorithm for Least-Squares Estimation of Nonlinear Parameters}",
