+++ paper.md
@@ -40,9 +40,9 @@ where obs is the observed spectrum, synth the synthetic, and &sigma; the error o
 
 The synthetic spectra, i.e. the fluxes at the top of the photosphere, are created on-the-fly with the radiative transfer code, [``MOOG``](https://www.as.utexas.edu/~chris/moog.html) [@Sneden:1973] in fortran for a set of stellar parameters. The synthetic spectra are later convolved with different rotational profiles, such as macroturbulence (vmac), projected rotational velocity (vsini), and instrumental broadening (resolution) to match the observations. To synthesize a spectrum, the information on how the physical properties (such as temperature, electron pressure, gas pressure, etc.) behave at each layer (optical depth) of the atmosphere is necessary. This information is provided in a tabulated form by interpolating from a grid of pre-computed stellar atmospheres for a set of stellar parameters (Teff, logg, [M/H]). ``FASMA`` includes two sets of grids for this analysis: Kurucz [@Kurucz:1993] or marcs models [@Gustafsson:2008].
 
-A key component for spectral synthesis is the list of atoms and molecules which are included in the wavelength intervals. The spectral line list required for the derivation of stellar parameters is described in @Tsantaki:2018 and for the chemical abundances is taken from @Adibekyan:2015.
+A key component for spectral synthesis is the list of atoms and molecules which are included in the wavelength intervals. The spectral line list required for the derivation of stellar parameters is described in @Tsantaki:2018 and for the chemical abundances is taken from @Adibekyan:2015 and @Delgado:2015.
 
-Before we compare with observations, we perform a local normalization for the regions of the synthesis and filter out cosmic rays. The procedure to derive stellar parameters is presented in detail in [@Tsantaki:2018] and has been tested for medium and high resolution spectrographs. In the new version 2.0, we now include a new feature to derive chemical abundances for the following elements: Na, Mg, Al, Si, Ca, Sc, Ti, V, Cr, Mn, and Ni. For this process, stellar parameters have to be set and the only free parameter is the abundance of each element. The abundances are calculated in a region of &pm;2m &angst; around each spectral line. The line list of the neighboring lines is taken from [VALD](http://vald.astro.uu.se/~vald/php/vald.php) [@Ryabchikova:2015].
+Before we compare with observations, we perform a local normalization for the regions of the synthesis and filter out cosmic rays. The procedure to derive stellar parameters is presented in detail in [@Tsantaki:2018] and has been tested for medium and high resolution spectrographs. In the new version 2.0, we now include a new feature to derive chemical abundances for the following elements: Li Na, Mg, Al, Si, Ca, Sc, Ti, V, Cr, Mn, and Ni. For this process, stellar parameters have to be set and the only free parameter is the abundance of each element. The abundances are calculated in a region of &pm;2m &angst; around each spectral line. The line list of the neighboring lines is taken from [VALD](http://vald.astro.uu.se/~vald/php/vald.php) [@Ryabchikova:2015].
 
 ``FASMA`` is run via terminal by setting the user options in a configuration file. ``FASMA`` includes all the standard inputs for spectral synthesis along with a manual for the derivation of stellar parameters and chemical abundances. The user has to provide solely the stellar spectrum for the analysis. ``FASMA`` can be used directly for most optical surveys, such as the Gaia-ESO [@Gilmore:2012] but also for the characterization planet hosts (see the [SWEET-cat](https://www.astro.up.pt/resources/sweet-cat/)).
 
