+++ FASMA/tests/test_interpolation.py
@@ -1,5 +1,6 @@
 from unittest import TestCase
 from FASMA.interpolation import solar_abundance
+from FASMA import FASMA
 
 
 class TestSolarAbundance(TestCase):
@@ -16,3 +17,12 @@ class TestSolarAbundance(TestCase):
         index, abundance = solar_abundance(element)
         self.assertTrue(index is None)
         self.assertTrue(abundance is None)
+
+
+class TestAbundances(TestCase):
+
+    elements = ['Li', 'Na', 'Mg', 'Al', 'Si', 'Ca', 'Sc', 'Ti', 'V', 'Cr', 'Mn', 'Ni']
+    for el in elements:
+        options = {'linelist':'../rawLinelist/elements.lst', 'intervals_file':'../rawLinelist/intervals_elements.lst', 'element':el, 'resolution':115000, 'observations': '../spectra/Sun_HARPS.fits'}
+        r = FASMA(**options)
+        self.assertTrue(r['abund'] is not None)
