+++ README.md
@@ -9,7 +9,7 @@ FASMA delivers the atmospheric stellar parameters (effective temperature, surfac
 
 The python code runs with python 3 and is wrapped around the spectral synthesis package in fortran: [MOOG version 2019](http://www.as.utexas.edu/~chris/moog.html) (Sneden et al. 1973).
 
-The spectral synthesis technique requires model atmospheres for the calculation of the synthetic spectra by MOOG which is provided with FASMA. FASMA includes two grids of models: Kurucz and marcs.
+The spectral synthesis technique requires model atmospheres for the calculation of the synthetic spectra by MOOG which is provided with FASMA. FASMA includes two grids of models in MOOG readable format, [Kurucz](http://research.iac.es/proyecto/ATLAS-APOGEE/) and [marcs](https://marcs.astro.uu.se/) that cover the parameter space for both dwarf and giant stars with metallicity limit of -5.0 dex.
 The instructions below are also included [here](https://github.com/MariaTsantaki/FASMA-synthesis/blob/master/manual/Manual_fasma.pdf).
 
 Contact here for bugs: tsantaki@arcetri.astro.it
@@ -17,14 +17,14 @@ Contact here for bugs: tsantaki@arcetri.astro.it
 # Installation
 Installing FASMA requires 2 simple steps.
 
-1) Run the installation script `install_fasma.sh`, in the FASMA-synthesis folder, to unzip the model atmospheres and install MOOG (MOOG uses gfortan, f77, or g77 compilers). The user will be asked about system requirements:
+**1)** Run the installation script `install_fasma.sh`, in the `FASMA-synthesis` folder, in order to unzip the model atmospheres and install MOOG (MOOG uses gfortan, f77, or g77 compilers). During this installation, the user will be asked about system requirements:
 "rh64" for 64-bit linux system, "rh" for 32-bit linux system, "maclap" for mac laptop, "macdesk" for mac desktop.
 
 ```sh
 $ ./install_fasma.sh
 ```
 
-2) Then install FASMA with pip:
+**2)** Then install FASMA with pip:
 ```sh
 $ pip install .
 ```
@@ -44,35 +44,44 @@ $ pip uninstall FASMA
 ```
 
 # Usage
+
+## FASMA by the configuration file
+
 FASMA is so easy. You can run FASMA from the terminal by configuring the options in the `config.yml` file and then run in the working directory:
 
 ```sh
 $ fasma
 ```
 
-If the configuration file is called something else (recommended for different projects), then run:
+If the configuration file is called something else (e.g. myConf.cfg), then run:
 
 ```sh
 $ fasma myConf.cfg
 ```
 
-FASMA includes a log file, `fasma.log`, which catches errors in order to inform the user.
+FASMA produces a log file, `fasma.log`, which catches errors in order to inform the user.
+
+## FASMA with scripts
 
 For large lists of stars, it is preferable to use FASMA inside scripts. The configuration options are inserted from a dictionary. The main function to use is FASMA().
 
 ```python
 FASMA(cfgfile='config.yml', overwrite=None, **kwargs)
 Input
------
+
+---
+
 cfgfile : str
   Configuration file (default: config.yml)
 overwrite : bool
   Overwrite the synthresults.dat file (default: False)
 
 Output
-------
+
+---
+
 synthresults.dat : file
-  Easy readable table with results
+  Easy readable table with results (tab separated)
 params : dict
   output parameters in dictionary
 ```
@@ -88,10 +97,10 @@ result = FASMA(**options)
 feh = result['feh']
 ```
 
-The result is a dictionary with the final parameters which are saved to a file (synthresults.dat or synthresults_elements.dat). The expected output dictionary of FASMA() includes the following:
[…]
