+++ docs/build/html/_sources/index.rst.txt
@@ -1,0 +1,20 @@
+.. FASMA documentation master file, created by
+   sphinx-quickstart on Mon Mar  2 18:04:07 2020.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+FASMA documentation
+===================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
