+++ docs/build/html/_sources/modules.rst.txt
@@ -1,0 +1,7 @@
+FASMA
+=====
+
+.. toctree::
+   :maxdepth: 4
+
+   FASMA
