+++ docs/FASMA.rst
@@ -1,15 +1,3 @@
-.. FASMA documentation master file, created by
-   sphinx-quickstart on Mon Mar  2 18:04:07 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-FASMA documentation
-===================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
 FASMA package
 =============
 
@@ -19,6 +7,9 @@ Subpackages
 .. toctree::
 
    FASMA.tests
+
+Submodules
+----------
 
 FASMA.fasma module
 ------------------
@@ -93,3 +84,10 @@ FASMA.utils module
    :show-inheritance:
 
 
+Module contents
+---------------
+
+.. automodule:: FASMA
+   :members:
+   :undoc-members:
+   :show-inheritance:
