+++ docs/index.rst
@@ -1,0 +1,21 @@
+.. FASMA documentation master file, created by
+   sphinx-quickstart on Mon Mar  2 18:43:29 2020.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to FASMA's documentation!
+=================================
+
+.. toctree::
+   :maxdepth: 4
+   :caption: Contents:
+
+   FASMA
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
