+++ docs/build/html/_sources/index.rst.txt
@@ -10,6 +10,99 @@ FASMA documentation
    :maxdepth: 2
    :caption: Contents:
 
+FASMA package
+=============
+
+Subpackages
+-----------
+
+.. toctree::
+
+   FASMA.tests
+
+Submodules
+----------
+
+FASMA.fasma module
+------------------
+
+.. automodule:: FASMA.fasma
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.interpolation module
+--------------------------
+
+.. automodule:: FASMA.interpolation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.minimization module
+-------------------------
+
+.. automodule:: FASMA.minimization
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.mpfit module
+------------------
+
+.. automodule:: FASMA.mpfit
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.observations module
+-------------------------
+
+.. automodule:: FASMA.observations
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.solar\_abundance module
+-----------------------------
+
+.. automodule:: FASMA.solar_abundance
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.synthDriver module
+------------------------
+
+.. automodule:: FASMA.synthDriver
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.synthetic module
+----------------------
+
+.. automodule:: FASMA.synthetic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.utils module
+------------------
+
+.. automodule:: FASMA.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: FASMA
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 
 Indices and tables
