+++ docs/index.rst
@@ -1,21 +1,95 @@
 .. FASMA documentation master file, created by
-   sphinx-quickstart on Mon Mar  2 18:43:29 2020.
+   sphinx-quickstart on Mon Mar  2 18:04:07 2020.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to FASMA's documentation!
-=================================
+FASMA documentation
+===================
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 2
    :caption: Contents:
 
-   fasma
+FASMA package
+=============
+
+Subpackages
+-----------
+
+.. toctree::
+
+   FASMA.tests
+
+FASMA.fasma module
+------------------
+
+.. automodule:: FASMA.fasma
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.interpolation module
+--------------------------
+
+.. automodule:: FASMA.interpolation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.minimization module
+-------------------------
+
+.. automodule:: FASMA.minimization
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.mpfit module
+------------------
+
+.. automodule:: FASMA.mpfit
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.observations module
+-------------------------
+
+.. automodule:: FASMA.observations
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.solar\_abundance module
+-----------------------------
+
+.. automodule:: FASMA.solar_abundance
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.synthDriver module
+------------------------
+
+.. automodule:: FASMA.synthDriver
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.synthetic module
+----------------------
+
+.. automodule:: FASMA.synthetic
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+FASMA.utils module
+------------------
+
+.. automodule:: FASMA.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
