+++ docs/conf.py
@@ -10,9 +10,9 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-# import os
-# import sys
-# sys.path.insert(0, '/export/home/elettra/tsantaki/FASMA-synthesis/FASMA')
+import os
+import sys
+sys.path.insert(0, '../..')
 
 
 # -- Project information -----------------------------------------------------
