+++ requirements.txt
@@ -1,6 +1,7 @@
 numpy>=1.7.0
 scipy>=0.16
 matplotlib
+tkinter
 PyYAML
 statsmodels
 patsy
