+++ requirements.txt
@@ -1,7 +1,7 @@
 numpy>=1.7.0
 scipy>=0.16
 matplotlib
-tkinter
+Tkinter
 PyYAML
 statsmodels
 patsy
