### 2020-03-02 ecb5290
remove greek char
