+++ paper.md
@@ -32,11 +32,11 @@ bibliography: paper.bib
 
 Effective temperature (Teff), surface gravity (logg), and metallicity ([M/H]) are basic stellar atmospheric parameters necessary to characterize a star. Once these parameters are obtained, we can in turn, infer their chemical abundances of various elements and in conjunction with evolutionary models to estimate their evolution, i.e. mass and radius. In this work, we use spectroscopy as a powerful tool to extract this information from stellar atmospheres applied to stars with spectral type FGK both dwarfs and giants. The growing number of spectroscopic surveys dedicated to the study of the Galactic stellar populations has inflated the number of high quality spectra to several hundreds of thousands. This amount is expected to multiply with the forthcoming surveys, such as WEAVE [@deJong:2019] and 4MOST [@Dalton:2014]. The success of these surveys highly depends on the analysis tools to exploit sufficiently all spectral information. Moreover, it is a well-known axiom in exoplanetary studies that one can only determine the planetary properties once the ones of the host star are known. The planetary properties such as mass, radius, composition, are directly linked to their hosts and therefore, robust tools for the derivation of these parameters are necessary.
 
-Our goal is to provide a tool to determine accurately and precisely the stellar parameters and chemical abundances to fulfill the above purposes. ``FASMA`` is a Python package to derive the main stellar atmospheric parameters based on the spectral synthesis technique. The principle of this technique relies on the comparison of synthetic spectra with observations to yield the best-fit parameters under a &chi;<sup>2</sup> minimization process:
+Our goal is to provide a tool to determine accurately and precisely the stellar parameters and chemical abundances to fulfill the above purposes. ``FASMA`` is a Python package to derive the main stellar atmospheric parameters based on the spectral synthesis technique. The principle of this technique relies on the comparison of synthetic spectra with observations to yield the best-fit parameters under a x<sup>2</sup> minimization process:
 
- &chi;<sup>2</sup> = &Sigma; (obs<sub>i</sub> - synth<sub>i</sub>)<sup>2</sup>/&sigma;<sub>i</sub>
+ x<sup>2</sup> = &Sigma; (obs<sub>i</sub> - synth<sub>i</sub>)<sup>2</sup>/sigma<sub>i</sub>
 
-where obs is the observed spectrum, synth the synthetic, and &sigma; the error on the observed flux for each spectral point, i. The best parameters are the ones that minimize the non-linear least-squares based on the Levenberg-Marquardt algorithm [@Marquardt:1963].
+where obs is the observed spectrum, synth the synthetic, and sigma the error on the observed flux for each spectral point, i. The best parameters are the ones that minimize the non-linear least-squares based on the Levenberg-Marquardt algorithm [@Marquardt:1963].
 
 The synthetic spectra, i.e. the fluxes at the top of the photosphere, are created on-the-fly with the radiative transfer code, [``MOOG``](https://www.as.utexas.edu/~chris/moog.html) [@Sneden:1973] in fortran for a set of stellar parameters. The synthetic spectra are later convolved with different rotational profiles, such as macroturbulence (vmac), projected rotational velocity (vsini), and instrumental broadening (resolution) to match the observations. To synthesize a spectrum, the information on how the physical properties (such as temperature, electron pressure, gas pressure, etc.) behave at each layer (optical depth) of the atmosphere is necessary. This information is provided in a tabulated form by interpolating from a grid of pre-computed stellar atmospheres for a set of stellar parameters (Teff, logg, [M/H]). ``FASMA`` includes two sets of grids for this analysis: Kurucz [@Kurucz:1993] or marcs models [@Gustafsson:2008].
 
