### 2020-03-02 a7a9fda
Delete Manual_fasma.tex
