+++ manual/Manual_fasma.tex
@@ -1,288 +1,0 @@
-\documentclass[a4paper,12pt]{article}
-%\usepackage[a4paper]{geometry}
-\usepackage[top=1in, bottom=1.25in, left=0.5in, right=0.5in]{geometry}
-\usepackage[utf8]{inputenc}
-\usepackage{listings}
-\usepackage{color}
-\usepackage{hyperref}
-\definecolor{backcolour}{rgb}{0.95,0.95,0.92}
-\definecolor{codegreen}{rgb}{0,0.6,0}
-\lstdefinestyle{mystyle}{
-    backgroundcolor=\color{backcolour},
-    keepspaces=true,
-    showspaces=false,
-    showstringspaces=false,
-    showtabs=false,
-    tabsize=2
-}
-
-\lstset{style=mystyle}
-%opening
-\title{FASMA tutorial}
-\author{}
-
-\begin{document}
-
-\maketitle
-
-
-\section{The configuration file}
-
-To use FASMA, configure the \texttt{config.cfg} which is in YAML format. The configuration file contains all the input parameters necessary for FASMA set by the user. 
-Because this format is whitespace sensitive, you can check if the format is correct here: http://www.yamllint.com/ . This file must be in the folder where FASMA runs. 
-There are various configuration combinations depending on what the user wants. A mandatory requirement is the line list and the interval file, where the whole path needs to be inserted. 
-{\bf The option Number 2 is recommended for the parameter determination.}
-
-\begin{enumerate}
-
- \item This is the simplest option for FASMA where a synthetic spectrum is created. The rest options are set to default.
- {\footnotesize
- \begin{lstlisting}[language=Python]
-#
-star:
-  linelist: /home/FASMA_synthesis/FASMA/rawLinelist/linelist.lst
-  intervals_file: /home/FASMA_synthesis/FASMA/rawLinelist/intervals.lst
-\end{lstlisting}
-}
-
-The default line list is in the \texttt{rawLinelist} folder already provided within FASMA.
- \item For a given line list select your options, e.g.:
- {\footnotesize
- \begin{lstlisting}[language=Python]
- #linelist options
-star:
-  linelist: /home/FASMA_synthesis/FASMA/rawLinelist/linelist.lst
-  intervals_file: /home/FASMA_synthesis/FASMA/rawLinelist/intervals.lst
-  fix_vmac: true
-  fix_vt: true
-  minimize: true
-  refine: true
-  observations: /home/FASMA_synthesis/FASMA/spectra/Sun_HARPS.fits
-  resolution: 115000
-  plot: true
-\end{lstlisting}}
-
-This example indicates that for the observed spectrum \texttt{Sun\_HARPS.fits} with resolution 115000, FASMA will start the minimization process starting from solar values, keeping the 
-parameters vmac and vt fixed and using the refine option (see below). This option was used to derive parameters in Tsantaki et al. (2018). The spectral regions for the synthesis are 
-difined in the intervals.lst file with the option intervals\_file. The final spectrum will be plotted. For the not listed options, FASMA uses the default ones. {\bf The complete paths 
-of the line list, spectrum, and intervals file should be given.} FASMA already includes the line list (\texttt{linelist.lst}), the interval file (\texttt{intervals.lst}) and a HARPS 
-spectrum of the Sun (\texttt{Sun\_HARPS.fits})
-
- \item For a given line list and a set of initial values, create a synthetic spectrum, save the output in the results folder and plot it. 
- {\footnotesize
-
-\begin{lstlisting}[language=Python]
- #linelist options
-star:
-  teff: 5777
-  logg: 4.44
-  vmac: 3.21
[…]
