### 2020-03-02 123a3af
Delete test_interpolation.py
