### 2020-03-02 #2048 @MariaTsantaki
@warrickball About installation:
I agree it would be better to control everything with pip but I did not succeed so I spit the installation in 2 steps as it was proposed.
