### 2020-03-02 #2048 @MariaTsantaki
@warrickball  About Performance: I added a comment for the users to prefer fits rather than text. Also I included instructions about the format of the input files.

About documentation: I understand that the information was split into two places (README and manual) so I merged them. There are a lot of things to improve (I wish I knew sphinx earlier). Now it seems very difficult to compile the current documentation to readthedocs (https://readthedocs.org/projects/fasma-synthesis/). I failed but it is something I want to look in the future. Nevertheless, I updated the documentation.

About Automated tests: I included more in the tests folder about the minimization process. This code is run separately from installation. This is the only way I could think of.

State of the field: There are indeed other packages such as SME (http://www.stsci.edu/~valenti/sme.html) and ispec (https://www.blancocuaresma.com/s/iSpec) that use the spectral synthesis.

References: In fact there is a license for mpfit. It was first written in idl where I have included the reference in the paper. https://pages.physics.wisc.edu/~craigm/idl/fitqa.html
