+++ paper.bib
@@ -242,6 +242,20 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@INBOOK{Markwardt:2009,
+       author = {{Markwardt}, C.~B.},
+        title = "{Non-linear Least-squares Fitting in IDL with MPFIT}",
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
+    booktitle = {Astronomical Data Analysis Software and Systems XVIII},
+         year = 2009,
+       editor = {{Bohlender}, D.~A. and {Durand}, D. and {Dowler}, P.},
+       volume = {411},
+       series = {Astronomical Society of the Pacific Conference Series},
+        pages = {251},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2009ASPC..411..251M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Marquardt:1963,
    author = {{Marquardt}, D.~W.},
     title = "{An Algorithm for Least-Squares Estimation of Nonlinear Parameters}",
