### 2020-03-03 #2048 @warrickball
Hi @MariaTsantaki, I can see you've made big improvements to the package!

* **Installation**: I'm happy with the two-step installation. I just re-installed FASMA from scratch without a problem.
* **Tests**: The new `test_FASMA.py` is a good start and good enough to satisfy me for this review. It's a bit difficult to make sense of all of the output but it ran to completion without any obvious errors so I'm happy with that.
* **Paper**: I'd still like the paper to mention other spectral synthesis packages and how FASMA differs. Otherwise, it looks fine, although some of the mathematical typesetting is incorrect (e.g. Teff, the χ² equation).
* **Documentation**: Thanks for combining the sources. The `README.md` alone is now fine for command-line use (after all, I managed!) and OK for the Python interface. I can see that you're trying to create Sphinx documentation, which is great. I'd say don't worry about fighting to get it working on ReadTheDocs just for this submission. I'm happy for you to just tell users they can go to `FASMA/docs` and build the documentation with `make html`. That said, the repo currently looks like you've ended up with two Sphinx setups, one with source in `docs/` and another with source in `docs/source`. I managed to generate documentation by changing `SOURCEDIR` from `source` to `.` in the `Makefile`. If you're having trouble, I'd suggest wiping `docs` and starting again. Also, only add to the repo the files that you need to generate the docs, not Sphinx's outputs.

So for me, the outstanding items are:
1. mentioning other packages in the paper (*State of the field*),
2. create community guidelines (*Community guidelines*), and
3. the licensing of MOOG (*License*).

@xuanxu, can you comment on item 3? I don't feel expert enough of licensing issues to know whether the current situation with MOOG is OK. MOOG is bundled but doesn't have an obvious license. Basically, I want someone more knowledgeable than me to say "This is fine" and I'll be happy.
