+++ paper.bib
@@ -139,6 +139,100 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Magrini:2013,
+       author = {{Magrini}, Laura and {Randich}, Sofia and {Friel}, Eileen and
+         {Spina}, Lorenzo and {Jacobson}, Heather and {Cantat-Gaudin}, Tristan and
+         {Donati}, Paolo and {Baglioni}, Roberto and {Maiorca}, Enrico and
+         {Bragaglia}, Angela and {Sordo}, Rosanna and {Vallenari}, Antonella},
+        title = "{FAMA: An automatic code for stellar parameter and abundance determination}",
+      journal = {\aap},
+     keywords = {methods: data analysis, techniques: spectroscopic, stars: abundances, Galaxy: abundances, surveys, open clusters and associations: general, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2013,
+        month = oct,
+       volume = {558},
+          eid = {A38},
+        pages = {A38},
+          doi = {10.1051/0004-6361/201321844},
+archivePrefix = {arXiv},
+       eprint = {1307.2367},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2013A&A...558A..38M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Andreasen:2017,
+       author = {{Andreasen}, D.~T. and {Sousa}, S.~G. and {Tsantaki}, M. and
+         {Teixeira}, G.~D.~C. and {Mortier}, A. and {Santos}, N.~C. and
+         {Su{\'a}rez-Andr{\'e}s}, L. and {Delgado-Mena}, E. and
+         {Ferreira}, A.~C.~S.},
+        title = "{SWEET-Cat update and FASMA. A new minimization procedure for stellar parameters using high-quality spectra}",
+      journal = {\aap},
+     keywords = {techniques: spectroscopic, stars: atmospheres, planets and satellites: fundamental parameters, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2017,
+        month = apr,
+       volume = {600},
+          eid = {A69},
+        pages = {A69},
+          doi = {10.1051/0004-6361/201629967},
+archivePrefix = {arXiv},
+       eprint = {1703.06671},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2017A&A...600A..69A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Tabernero:2019,
+       author = {{Tabernero}, H.~M. and {Marfil}, E. and {Montes}, D. and
+         {Gonz{\'a}lez Hern{\'a}ndez}, J.~I.},
+        title = "{STEPAR: an automatic code to infer stellar atmospheric parameters}",
+      journal = {\aap},
+     keywords = {techniques: spectroscopic, methods: data analysis, stars: fundamental parameters, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2019,
+        month = aug,
+       volume = {628},
+          eid = {A131},
+        pages = {A131},
+          doi = {10.1051/0004-6361/201935465},
+archivePrefix = {arXiv},
+       eprint = {1907.06512},
+ primaryClass = {astro-ph.SR},
[…]
