+++ paper.md
@@ -30,13 +30,15 @@ bibliography: paper.bib
 
 # Summary
 
-Effective temperature (Teff), surface gravity (logg), and metallicity ([M/H]) are basic stellar atmospheric parameters necessary to characterize a star. Once these parameters are obtained, we can in turn, infer their chemical abundances of various elements and in conjunction with evolutionary models to estimate their evolution, i.e. mass and radius. In this work, we use spectroscopy as a powerful tool to extract this information from stellar atmospheres applied to stars with spectral type FGK both dwarfs and giants. The growing number of spectroscopic surveys dedicated to the study of the Galactic stellar populations has inflated the number of high quality spectra to several hundreds of thousands. This amount is expected to multiply with the forthcoming surveys, such as WEAVE [@deJong:2019] and 4MOST [@Dalton:2014]. The success of these surveys highly depends on the analysis tools to exploit sufficiently all spectral information. Moreover, it is a well-known axiom in exoplanetary studies that one can only determine the planetary properties once the ones of the host star are known. The planetary properties such as mass, radius, composition, are directly linked to their hosts and therefore, robust tools for the derivation of these parameters are necessary. 
+Effective temperature (Teff), surface gravity (logg), and metallicity ([M/H]) are basic stellar atmospheric parameters necessary to characterize a star. Once these parameters are obtained, we can in turn, infer their chemical abundances of various elements and in conjunction with evolutionary models to estimate their evolution, i.e. mass and radius. In this work, we use spectroscopy as a powerful tool to extract this information from stellar atmospheres applied to stars with spectral type FGK both dwarfs and giants. The growing number of spectroscopic surveys dedicated to the study of the Galactic stellar populations has inflated the number of high quality spectra to several hundreds of thousands. This amount is expected to multiply with the forthcoming surveys, such as WEAVE [@deJong:2019] and 4MOST [@Dalton:2014]. The success of these surveys highly depends on the analysis tools to exploit sufficiently all spectral information. Moreover, it is a well-known axiom in exoplanetary studies that one can only determine the planetary properties once the ones of the host star are known. The planetary properties such as mass, radius, composition, are directly linked to their hosts and therefore, robust tools for the derivation of these parameters are necessary.
 
-Our goal is to provide a tool to determine accurately and precisely the stellar parameters and chemical abundances to fulfill the above purposes. ``FASMA`` is a Python package to derive the main stellar atmospheric parameters based on the spectral synthesis technique. The principle of this technique relies on the comparison of synthetic spectra with observations to yield the best-fit parameters under a x<sup>2</sup> minimization process:
+There are spectral packages in the literature based on different methods to derive the atmospheric parameters. A standard method for solar-like stars relies on the measurements of equivalent widths of iron lines and by imposing excitation and ionization equilibrium [@Magrini:2013; @Andreasen:2017; @Tabernero:2019]. Other methods are based on fitting synthetic spectra created on-the-fly with observations under a minimization procedure [@Valenti:1996; @Blanco:2014]. Each of the above methods has different limitations depending on the accuracy of the provided atomic data of the line lists, the resolution of the spectrographs, on the quality of the spectra (e.g.signal-to-noise) but also due to the star itself (e.g. rotation, spectral type).
 
- x<sup>2</sup> = &Sigma; (obs<sub>i</sub> - synth<sub>i</sub>)<sup>2</sup>/sigma<sub>i</sub>
[…]
