### 2020-03-03 #2048 @MariaTsantaki
@warrickball Thanks a lot! There is some documentation now. I added the 1st (paper) and 2nd (README) point.
