### 2020-03-05 #2048 @xuanxu
@warrickball Yes, we should address the MOOG license issue. If we don't know which license MOOG is using (I can't find it in their website) I suggest @MariaTsantaki or any other FASMA's author ask MOOG's author, maybe he can point us to where the license is or tell us if MOOG license allows redistributing. If that is not the case they could ask for explicit written permission for FASMA. Otherwise I guess we should assume copyright and FASMA could not bundle it in their code, adding it as a pre-requisite and including instructions about how to install it.
