### 2020-03-12 #2048 @MariaTsantaki
Hi all,
The author of MOOG is kindly ok with its redistribution. He mentioned to add a disclaimer on the README that we are not taking responsibility for proper use of MOOG as it is also mentioned in his website which I guess lies under the MIT license. I will ask him if possible to add the license online but he mentioned he is also willing to send a letter to the journal if necessary about this.
