### 2020-03-12 #2048 @xuanxu
@MariaTsantaki Great, if he specifies the MIT license online and/or adds [the license file](https://choosealicense.com/licenses/mit/) to the downloadable .tar.gz file with the code, that would be enough.
