### 2020-04-13 #2048 @arfon
> @arfon yeah, we're waiting for the author to address some minor issues mentioned by the reviewers to tick the last checkboxes, and then there's the license issue, I'm not sure if MOOG has already made explicit its license.

:+1: thanks!
