### 2020-04-21 #2048 @MariaTsantaki
Hi all,
The final point was about the license of MOOG that is available freely here: https://www.as.utexas.edu/~chris/moog.html. Before the pandemic, I had contacted the author who confirmed that MOOG is free to redistribute with the suggestion to add a disclaimer on the README that we are not taking responsibility for proper use of MOOG as it is also mentioned in his website but there is no explicit license or any official copyrights to the MOOG's website yet. Does this make it as in a 'public domain software' which means it is free for redistribution anyway?
I have added the relevant information about MOOG and its references in the paper, the README and in the MOOG folder.
Could you please confirm that these are enough to proceed with the paper?
Thank you!
Maria
