### 2020-04-22 #2048 @danielskatz
No license means JOSS can't review it - the copyright owner needs to specify an OSI-approved license
