### 2020-04-22 #2048 @ipelupessy
Is that correct? MOOG is a dependency of FASMA (which we are reviewing). If MOOG is really public domain that should not be a problem since the source is there. But I think we should try to clarify the licence situation with the author of MOOG - I think the problem  is a bit cultural/generational, I have seen this before with astronomy software that predates open source licensing. the authors do not see the need to add a license since they never done that before..
