### 2020-04-22 #2048 @danielskatz
Sorry, maybe I misunderstood.  What I am concerned with here is the license(s) of software in the repository that JOSS publishes, not those of dependencies in general.

Regarding "public domain", see https://opensource.org/node/878

Also note that JOSS is an OSI-affiliate, so the software it publishes must have OSI-approved licenses.
