### 2020-04-22 #2048 @xuanxu
@danielskatz the issue here is that a specific dependency (MOOG) is being redistributed entirely with this package, so its code is included in the FASMA repo. Author has granted permission but there's no license file in the[ original site](http://www.as.utexas.edu/~chris/moog.html), so the question is: Is [this file included in the MOOG folder inside FASMA](https://github.com/MariaTsantaki/FASMA-synthesis/blob/master/FASMA/MOOG/license.txt) enough for us to go proceed?
cc/@arfon
