### 2020-04-22 #2048 @danielskatz
The answer is no, unfortunately.

That file is kind of a statement about a license, but it's not really a license, nor is it an OSI-approved license.

This can most easily be resolved by the **author** of MOOG choosing an OSI-approved license, assuming he is the sole copyright owner and is allowed to do this.  No one but the copyright holder can set the license.
