### 2020-05-02 #2048 @ipelupessy
ouch, this is a pity! I think it would really be best the author to add a license to MOOG, the reasons for this are summarized e.g. here: https://blog.esciencecenter.nl/a-license-to-science-cd8030a4a145
the alternative is that he actually wants to dissallow further redistribution of the code (which is possible), but I am not sure if that is the case (I would be a bit surprised).
