### 2020-05-02 #2048 @MariaTsantaki
It is quite unfortunate.
@arfon @danielskatz just to confirm. Without an OSI-approved license for MOOG the process is stuck, right?
