### 2020-05-03 #2048 @arfon
Hi all, just getting up to speed here. I agree that JOSS can't publish this with MOOG bundled if MOOG hasn't licensed the code clearly with an OSI-approved license.

One option I don't see discussed here is moving the MOOG code out of this repository and having some kind of setup/Makefile that would download a copy of the MOOG source code for the end user. Is this something that might work?
