### 2020-05-06 #2048 @DanielAndreasen
Hi @arfon, I think that is an interesting idea. If we were to go down that road, it would look something like this:
1. Download MOOG from its home page.
2. Patch the code, in order to remove dependencies to SuperMongo, which the fewest have installed. This is similar to our current approach in FASMA.
3. Install the software.

Here I worry about the patching. Would that give us some licensing problems?
