### 2020-05-06 #2048 @arfon
> Here I worry about the patching. Would that give us some licensing problems?

With the standard caveat of me not being a lawyer and this not being legal advice, I think this is fine, especially if the approach we're talking about is for the end-user (i.e. the consumer of your `FASMA 2.0` software) running this download/patching command themselves. Is that the sort of thing you're thinking of?
