### 2020-05-06 #2048 @DanielAndreasen
Well, as stated by Chris Sneden (the author of MOOG), he has no problems with people editing and using his software (as long as he is not responsible for the results/conclusions drawn from it). So I don't think we are going to have any legal problems whatsoever.
If we were to go down this road, I would hope to make it as automatic as possible, that is, install MOOG and FASMA together with one command (possible two if needed).

I just don't really see the difference between:
* shipping FASMA with MOOG (current state)
* shipping FASMA with a script that download, edit, and install MOOG

It just looks like we are hiding MOOG behind a download/patch/install script.
