### 2020-05-06 #2048 @arfon
> I just don't really see the difference between:
> - shipping FASMA with MOOG (current state)

The issue here is ours. JOSS does not publish software without an OSI-approved license.

> Well, as stated by Chris Sneden (the author of MOOG), he has no problems with people editing and using his software (as long as he is not responsible for the results/conclusions drawn from it). So I don't think we are going to have any legal problems whatsoever.

I'm afraid it essentially doesn't matter what Chris Sneden has communicated to you directly. To make his position explicit, he should add a standard open source license which says as much. For example, the MIT license says:

> THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
