### 2020-05-06 #2048 @DanielAndreasen
I understand the issues here. But can we for example put MOOG in a separate github repo, and during the installation of FASMA, it will also install MOOG from that repository?
