### 2020-05-06 #2048 @arfon
Yes, I think that would be acceptable.
