### 2020-05-14 e16c85a
Remove MOOG dir
