+++ FASMA/MOOG/Params.f
@@ -1,813 +1,0 @@
-
-      subroutine params 
-c****************************************************************************** 
-c     This subroutine reads in the commands from the parameter file 
-c****************************************************************************** 
-
-      implicit real*8 (a-h,o-z) 
-      include 'Atmos.com' 
-      include 'Linex.com' 
-      include 'Factor.com' 
-      include 'Pstuff.com' 
-      include 'Obspars.com'
-      include 'Mol.com'
-      include 'Multistar.com'
-      real*8 deltalogab(5)
-      character keyword*20
-      character arrayz*80
-      integer kk
-      data newcount, linecount /0, 0/
-
-
-      if (linecount .eq. 0) oldcount = 0
-
-
-c  IF DOING MULTIPLE RUNS: if this is not the first reading of the
-c  parameter file, then read down to correct place in parameter file, 
-c  using "linecount", and then skip the re-initialization of the 
-c  various variables
-      rewind nfparam
-      read (nfparam,1001,end=100) arrayz
-      if (linecount .ne. 0) then
-         do i=1,linecount
-            read (nfparam,1001,end=100) arrayz
-         enddo
-         go to 4
-      endif
-
-
-c  INITIALIZE SOME VARIABLES: output file names and output file numbers
-c  f1out, nf1out:    verbose standard output
-c  f2out, nf2out:    raw synthetic spectra, or summary abundances, or
-c                    summary curves-of-growth
-c  f3out, nf3out:    smoothed synthetic spectra
-c  f4out, nf4out:    IRAF-text-style synthetic spectra output
-c  f5out, nf5out:    postscript plot output
-c  f6out, nf6out:    synthetic/observed comparison text output (gridsyn)
-c  f7out, nf7out:    raw synthetic spectrum of primary of a binary, 
-c                    output summary run table, or
-c                    to be used for some other purpose as needed; at
-c                    present not given keyword input
-c  f8out, nf8out:    raw synthetic spectrum of secondary of a binary,
-c                    or the kept lines in a line weedout, or for
-c                    some other purpose as needed
-c  f9out, nf9out:    combined smoothed synthetic spectrum of a binary star,
-c                    or the discarded lines in a line weedout, or the
-c                    mean raw synthesis for a population of stars, or
-c                    some other purpose as needed
-c                    to be used for some other purpose as needed`
-c  f10out, nf10out:  combined smoothed synthetic spectrum of a binary star,
-c                    to be used for some other purpose as needed`
-      f1out =   'no_filename_given'
-      f2out =   'no_filename_given'
-      f3out =   'no_filename_given'
-      f4out =   'no_filename_given'
-      f5out =   'optional_output_file'
-      f6out =   'optional_output_file'
-      f7out =   'no_filename_given'
-      f8out =   'no_filename_given'
-      f9out =   'no_filename_given'
-      f10out =  'no_filename_given'
-      nf1out =   0
-      nf2out =   0
-      nf3out =   0
-      nf4out =   0
-      nf5out =   0
-      nf6out =   0
-      nf7out =   0
-      nf8out =   0
-      nf9out =   0
-      nf10out =  0
-      modelnum = 0
-
-
-c  INITIALIZE SOME VARIABLES: input file names and input file numbers
-      fmodel =  'no_filename_given'
-      flines =  'no_filename_given'
-      fslines = 'no_filename_given'
-      fobs =    'no_filename_given'
-      ftable =  'no_filename_given' 
-      nfmodel =  0 
-      nflines =  0
-      nfslines = 0
-      nfobs =    0
-      nftable =  0
-
-
[…]
