+++ FASMA/MOOG/Tablepop.f
@@ -1,171 +1,0 @@
-
-      subroutine tablepop (option)
-c******************************************************************************
-c     this routine opens the table file containing information for a stellar
-c     population, and reads the data in that file; the information is a
-c     bit different for "abpop" and "synpop"
-c******************************************************************************
-
-      implicit real*8 (a-h,o-z)
-      include 'Atmos.com'
-      include 'Multimod.com'
-      include 'Pstuff.com'
-      integer option
-      character*80 line
-
-  
-c*****open the model table input file and the summary table output file
-      nftable = 18
-      lscreen = 4
-      array = 'MODEL TABLE INPUT FILE'
-      nchars = 22
-      call infile ('input ',nftable,'formatted  ',0,nchars,
-     .             ftable,lscreen)      
-      nf7out = 24
-      lscreen = 6
-      array = 'MODEL TABLE OUTPUT FILE'
-      nchars = 18
-      call infile ('output ',nf7out,'formatted  ',0,nchars,
-     .             f7out,lscreen)
-
-
-c*****read the table input for integrated light EW matching
-      if (option .eq. 1) then
-         read (nftable,1001) line
-         if (line(1:5) .ne. 'abpop' ) then
-            write(*,*) 'OOPS!  WRONG TABLE FOR ABPOP; I QUIT!'
-            stop
-         endif
-         do i=1,1000
-            read (nftable,1001) line
-            if     (line(1:5) .eq. 'modpr') then
-               call blankstring (modpre)
-               modpre(1:70) = line(11:80)
-            elseif (line(1:5) .eq. 'synpr') then
-               call blankstring (synpre)
-               synpre(1:70) = line(11:80)
-            elseif (line(1:5) .eq. 'title') then
-               call blankstring (abitle)
-               popitle(1:74) = line(7:80)
-               write (nf7out,1002) popitle(1:73)
-            elseif (line(1:5) .eq. 'model') then
-               do mmod=1,100
-                  if (mmod .eq. 100) then
-                     write(*,*) 'MORE THAN 99 MODELS; I QUIT!'
-                     stop
-                  endif
-                  read (nftable,*,end=10) j, radius(mmod), 
-     .                                    relcount(mmod)
-               enddo
-            endif
-         enddo
-
-
-c*****read the table input for integrated light spectrum syntheses
-      else
-         nabs = 0
-         nisos = 0
-         read (nftable,1001) line(1:80)
-         if (line(1:6) .ne. 'synpop') then
-            write (*,*) 'OOPS!  WRONG TABLE FOR SYNPOP; I QUIT!'
-            stop
-         endif
-         do k=1,1000
-            call blankstring (line)
-            read (nftable,1001) line(1:80)
-            if     (line(1:5) .eq. 'modpr') then
-               call blankstring (modpre)
-               modpre(1:70) = line(11:80)
-            elseif (line(1:5) .eq. 'synpr') then
-               call blankstring (synpre)
-               synpre(1:70) = line(11:80)
-            elseif (line(1:5) .eq. 'title') then
-               call blankstring (popitle)
-               popitle(1:74) = line(7:80)
-               write (nf7out,1003) popitle(1:74)
-            elseif (line(1:5) .eq. 'abund') then
-               read (line(12:80),*) nabs
-               if (nabs .gt. 0) then
-                  read (nftable,1001) line(1:80)
-                  read (line(1:80),*) (elspecial(i),i=1,nabs)
-                  write (nf7out,1009) (nint(elspecial(i)),i=1,nabs)
-               else
-                  write (nf7out,1010)
-               endif
-            elseif (line(1:5) .eq. 'isoto') then
-               read (line(10:80),*) nisos
-               if (nisos .gt. 0) then
[…]
