### 2020-05-17 #2048 @xuanxu
Thanks @MariaTsantaki!
@warrickball, @ipelupessy: Can you try the new installation method (it solves the License issue) and take a look at any pending items left from your checklist?
