### 2020-05-19 #2048 @warrickball
Hi,

I have very limited capacity at the moment but I still think I can turn in a review. I just tried the latest version and ran into a few issues.

When I ran `install_fasma.sh`, it got as far as starting to compile MOOG but failed with:

````Fortran
OpacHydrogen.f:163:50:

  163 |          call linter (wfflog,fflog(1,itheta),nnnn,wavelog,fftlog,1)
      |                                                  1
Error: Rank mismatch in argument ‘xnew’ at (1) (rank-1 and scalar)
OpacHydrogen.f:169:47:

  169 |      .              maxwave = map1(wbf,bf,nnnn,wave,hminbf,1)
      |                                               1
Error: Rank mismatch in argument ‘xnew’ at (1) (rank-1 and scalar)
make: *** [<builtin>: OpacHydrogen.o] Error 1
make: Leaving directory '/home/wball/work/FASMA-synthesis/FASMA/MOOG'
MOOGSILENT is not installed properly!
````
I'm compilng with gfortran 10.1.1 and suspected this is an issue with evolving Fortran standards, so I added the flag `-std=legacy` to `FASMA/MOOG/Makefile.rh64silent`, which allowed the build to finish. I also noticed that `install_fasma.sh` gave the warning
````
mv: cannot move 'MOOG' to 'FASMA/MOOG': Directory not empty
````
when I re-ran it, so you might want to add something like `rm -f FASMA/MOOG` to the install script.

What version(s) of gfortran have you been able to test, and what version(s) do you expect users to have? (Run `gfortran -v` to see what version is available in a given terminal session.) Given that I didn't hit this before, this issue might be unique to gfortran 10 but I think the `-std=legacy` flag was already available with earlier versions and would be compatible with any reasonably recent version.

The `pip` install was smooth and the tests ran to completion. I'd still like to try the simple examples I created with my own data but it'll be at least a few days before I have a chance to try them again. Aside from that, I think my comments on the paper have been addressed, so the only thing missing from the checklist is community guidelines.
