### 2020-05-20 #2048 @DanielAndreasen
Hi @warrickball,
Good you found a simple solution to the installation problem. It is worth including that in a FAQ (or even open and issue with your solution).

When I run `gfortran -v` I get a lot of information, but in the end it says `gcc version 7.5.0 (Ubuntu 7.5.0-3ubuntu1~18.04)`.
I will fix the small issue of removing the previous MOOG (in case one has to reinstall).
