### 2020-05-20 3b7516b
Remove old version of MOOG before a reinstall
