+++ install_fasma.sh
@@ -3,7 +3,10 @@
 tar -zxvf FASMA/models/apogee_kurucz.tar.gz -C FASMA/models/
 tar -zxvf FASMA/models/marcs.tar.gz -C FASMA/models/
 echo "Atmosphere models installed in dir: models"
-wget https://github.com/MariaTsantaki/moog-without-sm/archive/master.zip 
+
+# Installing MOOG
+rm -rf FASMA/MOOG
+wget https://github.com/MariaTsantaki/moog-without-sm/archive/master.zip
 unzip -q -o master.zip && mv moog-without-sm-master MOOG
 mv MOOG FASMA
 rm master.zip
