### 2020-05-24 #2048 @warrickball
I've managed to rerun the old example I had, so I'm happy with the state of the code. The top-level README does have a *Contributing* section so that enough for me to say my review is complete.
