### 2020-05-26 #2048 @xuanxu
Thanks @MariaTsantaki, everything looks good now, here are the next steps:

- Please release a [new tagged version](https://github.com/MariaTsantaki/FASMA-synthesis/releases) from the current master so it includes all the changes made during the review process
- Then archive that latest release in [Zenodo](https://zenodo.org/) (or any other similar service)
- Check the Zenodo deposit has the correct metadata: title and author name should match the paper; you may also add your ORCID.

Once you do that please report here the version number and archive DOI
