### 2020-05-26 #2048 @DanielAndreasen
While we are near the end of the review process, I would like to say thanks for a **very** transparent review. It has been such a pleasure working with a journal, where referees and editors **actively** improves upon the paper/software. I wish all journals could live up to this high standard.
