### 2020-06-08 #2048 @kyleniemeyer
Hi @MariaTsantaki, we are about ready to wrap this up, but in my review of the article I encountered some minor issues. Could you address these and let us know when you have done so?

- missing comma after "i.e." in the first and fourth paragraphs
- missing comma after the "e.g."s in the second paragraph
- Fortran should be capitalized
- the Tsantaki et al. 2018 citation should probably be changed from "... in (Tsantaki et al., 2018)" to "... by Tsantaki et al. (2018)", by changing `in [@Tsantaki:2018]` to `by @Tsantaki:2018`
- missing comma after "Li" in the list of elements on the second page
- it looks like a number of the references are missing the journal names: Adibekyan 2015, Andreasen 2017, Blanco-Cuaraesma 2014, Delgado 2015, Gustafsson 2008, Magrini 2013, Ryadchikova 2015, Sneden 1973, Tabernero 2019, Tsantaki 2018

Thanks!
