+++ paper.bib
@@ -119,8 +119,8 @@ archivePrefix = {arXiv},
          {Ridings}, Andy and {Guinouard}, Isabelle and {Verheijen}, Marc and
          {Tosh}, Ian and {Rogers}, Kevin and {Steele}, Iain and {Stuik}, Remko and
          {Tromp}, Neils and {Jasko}, Attila and {Kragt}, Jan and {Lesman}, Dirk and
-         {Mottram}, Chris and {Bates}, Stuart and {Gribbin}, Frank and {Fernand
-        o Rodriguez}, Luis and {Delgado}, Jos{\'e} M. and {Martin}, Carlos and
+         {Mottram}, Chris and {Bates}, Stuart and {Gribbin}, Frank and {Fernando
+         Rodriguez}, Luis and {Delgado}, Jos{\'e} M. and {Martin}, Carlos and
          {Cano}, Diego and {Navarro}, Ram{\'o}n and {Irwin}, Mike and
          {Lewis}, Jim and {Gonzalez Solares}, Eduardo and {O'Mahony}, Neil and
          {Bianco}, Andrea and {Zurita}, Christina and {ter Horst}, Rik and
@@ -140,63 +140,63 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Magrini:2013,
-       author = {{Magrini}, Laura and {Randich}, Sofia and {Friel}, Eileen and
-         {Spina}, Lorenzo and {Jacobson}, Heather and {Cantat-Gaudin}, Tristan and
-         {Donati}, Paolo and {Baglioni}, Roberto and {Maiorca}, Enrico and
-         {Bragaglia}, Angela and {Sordo}, Rosanna and {Vallenari}, Antonella},
-        title = "{FAMA: An automatic code for stellar parameter and abundance determination}",
-      journal = {\aap},
-     keywords = {methods: data analysis, techniques: spectroscopic, stars: abundances, Galaxy: abundances, surveys, open clusters and associations: general, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
-         year = 2013,
-        month = oct,
-       volume = {558},
-          eid = {A38},
-        pages = {A38},
-          doi = {10.1051/0004-6361/201321844},
-archivePrefix = {arXiv},
-       eprint = {1307.2367},
- primaryClass = {astro-ph.IM},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2013A&A...558A..38M},
+      author = {{Magrini}, Laura and {Randich}, Sofia and {Friel}, Eileen and
+      {Spina}, Lorenzo and {Jacobson}, Heather and {Cantat-Gaudin}, Tristan and
+      {Donati}, Paolo and {Baglioni}, Roberto and {Maiorca}, Enrico and
+      {Bragaglia}, Angela and {Sordo}, Rosanna and {Vallenari}, Antonella},
+      title = "{FAMA: An automatic code for stellar parameter and abundance determination}",
+      journal = {Astronomy & Astrophysics},
+      keywords = {methods: data analysis, techniques: spectroscopic, stars: abundances, Galaxy: abundances, surveys, open clusters and associations: general, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+      year = 2013,
+      month = oct,
+      volume = {558},
+      eid = {A38},
+      pages = {A38},
+      doi = {10.1051/0004-6361/201321844},
+      archivePrefix = {arXiv},
+      eprint = {1307.2367},
+      primaryClass = {astro-ph.IM},
[…]
