+++ paper.md
@@ -46,7 +46,7 @@ The synthetic spectra, i.e., the fluxes at the top of the photosphere, are creat
 
 A key component for spectral synthesis is the list of atoms and molecules which are included in the wavelength intervals. The spectral line list required for the derivation of stellar parameters is described by @Tsantaki:2018 and for the chemical abundances is taken from @Adibekyan:2015 and @Delgado:2015.
 
-Before we compare with observations, we perform a local normalization for the regions of the synthesis and filter out cosmic rays. The procedure to derive stellar parameters is presented in detail by [@Tsantaki:2018] and has been tested for medium and high resolution spectrographs. In the new version 2.0, we now include a new feature to derive chemical abundances for the following elements: Li, Na, Mg, Al, Si, Ca, Sc, Ti, V, Cr, Mn, and Ni. For this process, stellar parameters have to be set and the only free parameter is the abundance of each element. The abundances are calculated in a region of &pm;2m &angst; around each spectral line. The line list of the neighboring lines is taken from [VALD](http://vald.astro.uu.se/~vald/php/vald.php) [@Ryabchikova:2015].
+Before we compare with observations, we perform a local normalization for the regions of the synthesis and filter out cosmic rays. The procedure to derive stellar parameters is presented in detail by @Tsantaki:2018 and has been tested for medium and high resolution spectrographs. In the new version 2.0, we now include a new feature to derive chemical abundances for the following elements: Li, Na, Mg, Al, Si, Ca, Sc, Ti, V, Cr, Mn, and Ni. For this process, stellar parameters have to be set and the only free parameter is the abundance of each element. The abundances are calculated in a region of &pm;2m &angst; around each spectral line. The line list of the neighboring lines is taken from [VALD](http://vald.astro.uu.se/~vald/php/vald.php) [@Ryabchikova:2015].
 
 ``FASMA`` is run via terminal by setting the user options in a configuration file. ``FASMA`` includes all the standard inputs for spectral synthesis along with a manual for the derivation of stellar parameters and chemical abundances. The user has to provide solely the stellar spectrum for the analysis. ``FASMA`` can be used directly for most optical surveys, such as the Gaia-ESO [@Gilmore:2012] but also for the characterization planet hosts (see the [SWEET-cat](https://www.astro.up.pt/resources/sweet-cat/)).
 
