+++ paper.md
@@ -42,7 +42,7 @@ Our goal is to provide a tool to determine accurately and precisely the stellar 
 
 where obs is the observed spectrum, synth the synthetic, and $\sigma$ the error on the observed flux for each spectral point, i. The best parameters are the ones that minimize the non-linear least-squares based on the Levenberg-Marquardt algorithm [@Marquardt:1963; @Markwardt:2009].
 
-The synthetic spectra, i.e., the fluxes at the top of the photosphere, are created on-the-fly with the radiative transfer code, [``MOOG``](https://www.as.utexas.edu/~chris/moog.html) [@Sneden:1973] in Fortran for a set of stellar parameters. The synthetic spectra are later convolved with different rotational profiles, such as macroturbulence (vmac), projected rotational velocity (vsini), and instrumental broadening (resolution) to match the observations. To synthesize a spectrum, the information on how the physical properties (such as temperature, electron pressure, gas pressure, etc.) behave at each layer (optical depth) of the atmosphere is necessary. This information is provided in a tabulated form by interpolating from a grid of pre-computed stellar atmospheres for a set of stellar parameters (Teff, logg, [M/H]). ``FASMA`` includes two sets of grids for this analysis: Kurucz [@Kurucz:1993] or marcs models [@Gustafsson:2008].
+The synthetic spectra, i.e., the fluxes at the top of the photosphere, are created on-the-fly with the radiative transfer code, [``MOOG``](https://www.as.utexas.edu/~chris/moog.html) [@Sneden:1973] in Fortran for a set of stellar parameters. The synthetic spectra are later convolved with different rotational profiles, such as macroturbulence (vmac), projected rotational velocity (vsini), and instrumental broadening (resolution) to match the observations. To synthesize a spectrum, the information on how the physical properties (such as temperature, electron pressure, gas pressure, etc.) behave at each layer (optical depth) of the atmosphere is necessary. This information is provided in a tabulated form by interpolating from a grid of pre-computed stellar atmospheres for a set of stellar parameters (Teff, logg, [M/H]). ``FASMA`` includes two sets of grids for this analysis: Kurucz [@Kurucz:1993] or MARCS models [@Gustafsson:2008].
 
 A key component for spectral synthesis is the list of atoms and molecules which are included in the wavelength intervals. The spectral line list required for the derivation of stellar parameters is described by @Tsantaki:2018 and for the chemical abundances is taken from @Adibekyan:2015 and @Delgado:2015.
 
@@ -57,6 +57,6 @@ Before we compare with observations, we perform a local normalization for the re
 
 # Acknowledgements
 
-We acknowledge contributions from V. Adibekyan, H. Tabernero, and E. Delgado-Mena. We also thank the referees for improving this work. This research made use of the Vienna Atomic Line Database operated at Uppsala University, the Institute of Astronomy RAS in Moscow, and the University of Vienna. We thank the PyAstronomy and Astropy communities.
+We acknowledge contributions from V. Adibekyan, H. Tabernero, and E. Delgado-Mena. We also thank the referees for improving this work. This research made use of the Vienna Atomic Line Database operated at Uppsala University, the Institute of Astronomy RAS in Moscow, and the University of Vienna. This research made use of  PyAstronomy and [Astropy](http://www.astropy.org) communities [@astropy:2013, @astropy:2018].
 
 # References
