+++ paper.bib
@@ -470,97 +470,3 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2012Msngr.147...25G},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
-@ARTICLE{astropy:2018,
-       author = {{Astropy Collaboration} and {Robitaille}, Thomas P. and
-         {Tollerud}, Erik J. and {Greenfield}, Perry and {Droettboom}, Michael and
-         {Bray}, Erik and {Aldcroft}, Tom and {Davis}, Matt and
-         {Ginsburg}, Adam and {Price-Whelan}, Adrian M. and
-         {Kerzendorf}, Wolfgang E. and {Conley}, Alexander and {Crighton}, Neil and
-         {Barbary}, Kyle and {Muna}, Demitri and {Ferguson}, Henry and
-         {Grollier}, Fr{\'e}d{\'e}ric and {Parikh}, Madhura M. and
-         {Nair}, Prasanth H. and {Unther}, Hans M. and {Deil}, Christoph and
-         {Woillez}, Julien and {Conseil}, Simon and {Kramer}, Roban and
-         {Turner}, James E.~H. and {Singer}, Leo and {Fox}, Ryan and
-         {Weaver}, Benjamin A. and {Zabalza}, Victor and {Edwards}, Zachary I. and
-         {Azalee Bostroem}, K. and {Burke}, D.~J. and {Casey}, Andrew R. and
-         {Crawford}, Steven M. and {Dencheva}, Nadia and {Ely}, Justin and
-         {Jenness}, Tim and {Labrie}, Kathleen and {Lim}, Pey Lian and
-         {Pierfederici}, Francesco and {Pontzen}, Andrew and {Ptak}, Andy and
-         {Refsdal}, Brian and {Servillat}, Mathieu and {Streicher}, Ole},
-        title = "{Astropy: A community Python package for astronomy}",
-      journal = {Astronomy & Astrophysics},
-     keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools, Astrophysics - Instrumentation and Methods for Astrophysics},
-         year = 2013,
-        month = oct,
-       volume = {558},
-          eid = {A33},
-        pages = {A33},
-          doi = {10.1051/0004-6361/201322068},
-archivePrefix = {arXiv},
-       eprint = {1307.6212},
- primaryClass = {astro-ph.IM},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2013A&A...558A..33A},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{astropy:2018,
-       author = {{Astropy Collaboration} and {Price-Whelan}, A.~M. and
-         {Sip{\H{o}}cz}, B.~M. and {G{\"u}nther}, H.~M. and {Lim}, P.~L. and
-         {Crawford}, S.~M. and {Conseil}, S. and {Shupe}, D.~L. and
-         {Craig}, M.~W. and {Dencheva}, N. and {Ginsburg}, A. and {Vand
-        erPlas}, J.~T. and {Bradley}, L.~D. and {P{\'e}rez-Su{\'a}rez}, D. and
-         {de Val-Borro}, M. and {Aldcroft}, T.~L. and {Cruz}, K.~L. and
-         {Robitaille}, T.~P. and {Tollerud}, E.~J. and {Ardelean}, C. and
-         {Babej}, T. and {Bach}, Y.~P. and {Bachetti}, M. and {Bakanov}, A.~V. and
[…]
