+++ paper.bib
@@ -471,7 +471,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{astropy:2018,
+@ARTICLE{astropy:2013,
        author = {{Astropy Collaboration} and {Robitaille}, Thomas P. and
          {Tollerud}, Erik J. and {Greenfield}, Perry and {Droettboom}, Michael and
          {Bray}, Erik and {Aldcroft}, Tom and {Davis}, Matt and
